% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessibility_methods.R
\name{accessibility_add_intervals}
\alias{accessibility_add_intervals}
\title{Accessibility of amenities by travel time, to one dataframe}
\usage{
accessibility_add_intervals(
  point_per_DA,
  region_DA_IDs = region_DA_IDs,
  traveltimes,
  time_intervals = which(1:60\%\%5 == 0)
)
}
\arguments{
\item{point_per_DA}{A dataframe containing the number of data points per DA.
The first two columns should be \code{ID} and \code{geometry}, and the remaining
columns should be the number of points per amenity type.}

\item{region_DA_IDs}{<\verb{character vector}> All the current census'
DA IDs present in the region. Only those will be extracted from the database.}

\item{traveltimes}{A list of dataframes of travel times by mode. The output
of \code{\link[cc.buildr]{accessibility_get_travel_times}}.}

\item{time_intervals}{A vector with the time intervals to use (in
minutes). Defaults to \code{c(5, 10, 15, 20, 25, 30, 35, 40, 45, 50, 55, 60)}.}
}
\value{
A dataframe containing the number of accessible amenities per
DA in different time intervals, for different modes of transport.
}
\description{
This function adds the data from the point layer to the travel time
matrices, and summarizes the amount of reachable amenities in certain
time intervals.
}
