% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessibility_methods.R
\name{accessibility_get_travel_times}
\alias{accessibility_get_travel_times}
\title{Get travel time dataframes}
\usage{
accessibility_get_travel_times(
  modes = c("foot", "bike", "car", "transit"),
  region_DA_IDs
)
}
\arguments{
\item{modes}{<\verb{character vector}> Modes for which to download the travel
times}

\item{region_DA_IDs}{<\verb{character vector}> All the current census'
DA IDs present in the region. Only those will be extracted from the database.}
}
\value{
Returns a named list. Each mode is an element, and inside each
DA has its own dataframe with time to get to each other DA.
}
\description{
Get travel time dataframes
}
