% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessibility_methods.R
\name{accessibility_point_per_DA}
\alias{accessibility_point_per_DA}
\title{Get number of point per DA}
\usage{
accessibility_point_per_DA(point_data, DA_table, crs)
}
\arguments{
\item{point_data}{<\verb{named list}> List of \code{sf} dataframes. The only thing
that's used is the spatial features of each element of the list. Spatial
features must be points. The names of each element of the list is the name
of the variables.}

\item{DA_table}{A \code{DA} sf data.frame which will be used to intersect
each element of the \code{point_data} list.}

\item{crs}{<\code{numeric}> EPSG coordinate reference system to be assigned, e.g.
\code{32617} for Toronto.}
}
\value{
Outputs a single dataframe with the same number of rows as there
are in \code{DA_table}. Every element of the \code{point_data} list is a column in the
output, and the name of the column is the name of each element of the
\code{point_data} list.
}
\description{
Takes a named list of \code{sf} points. The name of the list is the name of the
variable. We use this function to know how many \code{X} are in each DA.
}
