% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_table.R
\name{add_high_corr_combination}
\alias{add_high_corr_combination}
\title{Add high correlation combinations to modules object}
\usage{
add_high_corr_combination(svm)
}
\arguments{
\item{svm}{<\code{list}> scales_variables_modules list object.}
}
\value{
<\code{list}> The modified svm object with an added \code{high_corr_combination} list
to each module within the \code{modules} data frame. This list contains names of
variables in \code{var_right} that have a high correlation (> 0.7) with variables in
\code{var_left}, for the chosen scale.
}
\description{
This function calculates correlations between variables defined in the
\code{var_left} and \code{var_right} of each module within the \code{modules} dataset of the
provided svm object. It then updates the svm object with these high correlation
combinations (correlation > 0.7) for dropdown styling. The function
specifically targets datasets located within specified scales and excludes
"building" scale when multiple scales are available. It utilizes \code{future.apply}
for parallel computation of correlations, aiming to enhance performance.
}
