% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{add_q3}
\alias{add_q3}
\title{Append q3 column to the dataframe}
\usage{
add_q3(df, vars, time_regex = "\\\\d{4}")
}
\arguments{
\item{df}{<\code{data.frame}> Contains all columns in \code{vars}.}

\item{vars}{<\verb{vector of character}> Contains all variable names from which
to add q3s. Must fit with a name in \code{df}.}

\item{time_regex}{<\code{character}> Regular expression which corresponds to
a timeframe, placed at the end of the \code{vars} vector. e.g. \code{"\\d{4}"} for
years.}
}
\value{
Returns the same data.frame as df with q3 columns appended.
}
\description{
Append q3 column to the dataframe
}
