% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{add_q5}
\alias{add_q5}
\title{Append q5 column to a dataframe}
\usage{
add_q5(df, breaks, time_regex = "_\\\\d{4}$", breaks_base)
}
\arguments{
\item{df}{<\code{data.frame}> Contains all columns in \code{vars}.}

\item{breaks}{<\code{data.frame}> A data.frame containing all the q5 breaks. The
output of \code{\link[cc.buildr]{get_breaks_q5}}.}

\item{time_regex}{<\code{character}> Regular expression which corresponds to
a timeframe, placed at the end of the \code{vars} vector. e.g. \verb{_\\\\d\{4\}$} for
years.}

\item{breaks_base}{<\verb{names list}> If we want variables to use the breaks of
another variable, we can supply a named list. The name is the variable that will
use the breaks of the value variable. ex. in access, we want all variables of
same theme to use the same breaks, no matter the transportation time. This
would be how we would do the list, if we want to use the 30 minutes break as
the basis for all the other transport time.
\code{list(access_transit_pwd_25_food_misc_2023 = "access_transit_pwd_30_food_misc_2023", access_transit_pwd_20_food_misc_2023 = "access_transit_pwd_30_food_misc_2023", ...)}}
}
\value{
Returns the same data.frame as df with q5 columns appended.
}
\description{
Append q5 column to a dataframe
}
