% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{add_regions_to_scales_dictionary}
\alias{add_regions_to_scales_dictionary}
\title{Add regions to scales dictionary}
\usage{
add_regions_to_scales_dictionary(
  scales_dictionary,
  regions,
  scales_consolidated,
  known_regions = NULL
)
}
\arguments{
\item{scales_dictionary}{<\code{data.frame}> The scales dictionary built using
\code{\link{build_census_scales}}.}

\item{regions}{<\verb{named list}> A list containing unioned regions, output of
\code{\link{create_master_polygon}}$regions.}

\item{scales_consolidated}{<\verb{named list}> A named list of sf data.frame
containing all scales listed with their regions. The output of
\code{\link[cc.buildr]{consolidate_scales}}.}

\item{known_regions}{<\verb{named list}> A named list of scales we already know
exactly which regions they should be assigned, useful for smaller scales requiring
a lot of computational power. e.g. \verb{list(grid250 = c("island", "city"),  grid100 = c("island", "city")}.
Defaults to NULL.}
}
\value{
A modified scales_dictionary with region information added.
}
\description{
This function adds region information to a given scales dictionary based on
spatial features and known regions. It filters out water features from Open
Street Map (OSM) and calculates the matching regions for each scale.
}
