% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_table.R
\name{add_var_right}
\alias{add_var_right}
\title{Add right variables to Modules}
\usage{
add_var_right(svm, ignore_to_choose_vr = c("^grd\\\\d", "^DB$", "^building$"))
}
\arguments{
\item{svm}{<\code{list}> scales_variables_modules list object.}

\item{exclude_scales}{<\verb{character vector}> Specifies the scales to be excluded from
the selection process of right variables. This is useful in scenarios where the
absence of data for certain scales (e.g., buildings not containing specific data)
should not affect the determination of right variables for a module. For instance,
if a module's autozoom feature displays buildings with DA data, scales related to
buildings without data can be excluded to ensure relevant right variable selection.}
}
\value{
<\code{list}> The modified \code{svm} object with \code{var_right} added to each
module in the \code{modules} dataframe.
}
\description{
This function iterates over the modules in a \code{scales_variables_modules$modules} object,
identifying and adding the appropriate right variables based on the classification
of left-side variables. It ensures that all right-side variable are valid and relevant.
}
