% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{addresses_db_links}
\alias{addresses_db_links}
\title{Provinces links to the National Open Database of Addresses}
\format{
\subsection{\code{addresses_db_links}}{

A data frame of 3 columns and the same number of rows as there are .zip of
addresses available
\describe{
\item{province_code}{The two character code of the Province}
\item{province}{The province name}
\item{link}{The download link to get the province's databse of addresses}
}
}
}
\source{
\url{https://www.statcan.gc.ca/en/lode/databases/oda}
}
\usage{
addresses_db_links
}
\description{
Provinces links to the National Open Database of Addresses
}
\keyword{datasets}
