% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_DA_ID.R
\name{append_DA_ID}
\alias{append_DA_ID}
\title{Append a DA ID to a data frame}
\usage{
append_DA_ID(DA_table, df, crs)
}
\arguments{
\item{DA_table}{<\verb{sf dataframe}> A spatial data frame representing a DA layer. The data
frame must contain an ID column that uniquely identifies each DA.}

\item{df}{<\verb{sf dataframe}> A spatial data frame with the spatial features for
which the district ID needs to be computed. The data frame must contain at
least one spatial column (geometry column).}

\item{crs}{<\code{character}> The coordinate reference system of both the district
layer and the data frame. This can be either a proj4 string or an epsg code.}
}
\value{
The data frame \code{df} with a new column called \code{DA_ID} that
contains the DA ID for each spatial feature in the data frame.
}
\description{
This function takes a spatial data frame of DAs , a data
frame with spatial information, and a coordinate reference system (CRS) for both
layers. The function then computes the DA ID of each spatial feature in
the data frame by checking if its centroid falls within a DA polygon. The DA
ID is appended to the data frame as a new column and the modified data frame is returned.
}
