% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_DA_popw_centroids.R
\name{append_DA_popw_centroids}
\alias{append_DA_popw_centroids}
\title{Append the population-weighted centroids to a DA table}
\usage{
append_DA_popw_centroids(DA_table, name = "popw_centroids_coords")
}
\arguments{
\item{DA_table}{<\code{data.frame}> A DA data.frame for which to merge IDs with the
\code{susbuildr::da_centroids_popw} data.}

\item{name}{<\code{character}> The name of the new column. Defaults to
\code{"popw_centroids_coords"}.}
}
\value{
The same table that was populated comes out with a new list column
containing the population-weighted centroids for each DA.
}
\description{
Append the population-weighted centroids to a DA table
}
