% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_empty_modules_table.R
\name{append_empty_modules_table}
\alias{append_empty_modules_table}
\title{Append an empty modules table to the scales list}
\usage{
append_empty_modules_table(scales)
}
\arguments{
\item{scales}{<\verb{named list}> A named list containings the scales and the
empty variables table. The output of
\code{\link[susbuildr]{append_empty_variables_table}}}
}
\value{
A list with the second index being an empty variables table.
}
\description{
\code{append_empty_modules_table()} takes the output of
\code{\link[susbuildr]{append_empty_variables_table}} and appends an empty
modules table.
}
