% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables_table.R
\name{append_empty_variables_table}
\alias{append_empty_variables_table}
\title{Append an empty variables table to the scales list}
\usage{
append_empty_variables_table(scales_consolidated)
}
\arguments{
\item{scales_consolidated}{<\verb{named list}> A named list of sf data.frame
containing all scales listed with their geos. The output of
\code{\link[susbuildr]{consolidate_scales}}.}
}
\value{
A list with the second index being an empty variables table.
}
\description{
\code{append_empty_variables_table()} adds a first level to the scales list, and
places an empty variables table to the scales list.
}
