% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_scale_to_dictionary.R
\name{append_scale_to_dictionary}
\alias{append_scale_to_dictionary}
\title{Add a scale to the scales dictionary}
\usage{
append_scale_to_dictionary(
  scales_dictionary,
  scale,
  sing,
  sing_with_article,
  plur,
  slider_title,
  place_heading,
  place_name,
  subtext
)
}
\arguments{
\item{scales_dictionary}{<\code{data.frame}> An already built scales dictionary
data.frame, most probably built using \code{\link[cc.buildr]{census_scales_dictionary}}.}

\item{scale}{<\code{character}> The scale code name, e.g. \code{"CSD"}.}

\item{sing}{<\code{character}> The scale name in singular, e.g. \code{"city"}.}

\item{plur}{<\code{character}> The scale name in plurial, e.g. \code{"cities"}.}

\item{slider_title}{<\code{character}> The scale name used on the map slider,
e.g. \code{"City"}. Can't start with a digit}

\item{place_heading}{<\code{character}> The place heading on a selection, e.g.
for "City of Laval" -> \code{"City of {name}"}.}

\item{place_name}{<\code{character}> The place name on a selection as the first
string of the dynamically generated text in the explore panel. e.g. for
"Laval has a population of ..." -> \code{"{name}"}.}

\item{subtext}{<\code{character}> Further quick and simple explanation of the scale
which will be placed in the scale dropdown. e.g. for \code{city}, \verb{Census scale (Census subdivision)}.}
}
\value{
The \code{scale_dictionary} data.frame that is fed with the
additional row.
}
\description{
Add a scale to the scales dictionary
}
