% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{aws_cli_cmd}
\alias{aws_cli_cmd}
\title{Execute AWS CLI Command and Parse JSON Output}
\usage{
aws_cli_cmd(call, aws_cli_profile = Sys.getenv("AWS_ADMIN_PROFILE"))
}
\arguments{
\item{call}{<\code{character}> The AWS CLI command to execute.}

\item{aws_cli_profile}{<\code{character}> The AWS CLI profile to use for the command.
Default is the value of the \code{AWS_ADMIN_PROFILE} environment variable.}
}
\value{
<\code{list}|\code{character}> Parsed JSON output from the executed AWS CLI command
or error message if an error occurred.
}
\description{
Executes a specified AWS CLI command using a given AWS CLI profile (or a default
if not specified). If successful, returns the parsed JSON output. If an error
occurs in the system call or JSON parsing, prints the error message instead.
}
