% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heroku.R
\name{aws_deploy}
\alias{aws_deploy}
\title{Deploy an app in ECR}
\usage{
aws_deploy(
  app_name,
  curbcut_branch = "HEAD",
  wd = getwd(),
  GA = FALSE,
  task_name = "CurbcutDeploy",
  aws_cli_profile,
  tag
)
}
\arguments{
\item{app_name}{<\code{character}> Must correspond to an existing ECR repository on
AWS account the user will login.}

\item{curbcut_branch}{<\code{character}> Is there a specific branch of Curbcut
that should be installed?}

\item{wd}{<\code{character}> Directory of the Curbcut-city repo.}

\item{GA}{<\code{logical}> Should google analytics be pushed with the image? Defaults
to FALSE. The function will also suggest to turn it to TRUE if the app name
is cc-montreal.}

\item{task_name}{<\code{character}> The name of the Windows task to build the Docker
container and upload to ECR AWS.}

\item{aws_cli_profile}{<\code{character}> The AWS IAM profile which should be used
to upload the container to ECR. Must have the right permissions. Must have
been previously set through the AWS CLI using \verb{aws configure --profile X}.}

\item{tag}{<\code{character}> What should be the tag of the ECR image? Use \code{latest}
to make it the most up-to-date image of the repository.}
}
\value{
Opens a terminal and disconnect from the current R session.
}
\description{
The function initiates the AWS ECR process and authenticates the
container. After successful authentication, it pushes the web application
to the specified ECR repo.
}
