% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{aws_duplicate_all}
\alias{aws_duplicate_all}
\title{Duplicate All AWS Components for a New City}
\usage{
aws_duplicate_all(new_city, from = "montreal")
}
\arguments{
\item{new_city}{<\code{character}> The name of the new city for which to duplicate all components.}

\item{from}{<\code{character}> The name of the city from which to duplicate. Optional.}
}
\value{
<\code{character}> A message indicating the outcome of the duplication process.
}
\description{
A wrapper function that duplicates all necessary AWS components for launching
in a new city. This includes task definitions, load balancers, target groups,
and ECS services.
}
