% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{aws_duplicate_LB}
\alias{aws_duplicate_LB}
\title{Duplicate AWS ELB and Target Groups for a New City}
\usage{
aws_duplicate_LB(new_city, from = "montreal")
}
\arguments{
\item{new_city}{<\code{character}> The name of the new city for which to duplicate
the load balancer and target groups.}

\item{from}{<\code{character}> The name of the city from which to duplicate.
Defaults to "montreal".}
}
\value{
<\code{character}> A message indicating the outcome of the duplication process.
}
\description{
Duplicates an existing AWS Elastic Load Balancer and its associated target groups,
adjusting names and attributes for a new city. It also handles the duplication of
listeners for both HTTP and HTTPS protocols.
}
