% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{aws_duplicate_service}
\alias{aws_duplicate_service}
\title{Duplicate an AWS ECS Service for a New City}
\usage{
aws_duplicate_service(new_city, from = "montreal")
}
\arguments{
\item{new_city}{<\code{character}> The name of the new city for which to duplicate the service.}

\item{from}{<\code{character}> The name of the city from which to duplicate.
Defaults to "montreal".}
}
\value{
<\code{character}> A message indicating the outcome of the service duplication process.
}
\description{
Duplicates an existing AWS ECS service for a new city, including updating the
task definition, target groups, and auto-scaling policies with city-specific names
and configurations.
}
