% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{aws_duplicate_task}
\alias{aws_duplicate_task}
\title{Duplicate an AWS ECS Task Definition for a new city}
\usage{
aws_duplicate_task(new_city, from = "montreal")
}
\arguments{
\item{new_city}{<\code{character}> The name of the new city for which to duplicate the task.}

\item{from}{<\code{character}> The name of the city from which to duplicate the task.
Defaults to "montreal".}
}
\value{
<\code{character}> A message indicating the outcome of the task duplication process.
}
\description{
Duplicates an existing AWS ECS task definition, substituting the specified city
name in the task definition, and registers it under the new city's name.
}
