% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessibility.R
\name{ba_accessibility_points}
\alias{ba_accessibility_points}
\title{Build and append pre-processed Access to Amenities}
\usage{
ba_accessibility_points(
  scales_variables_modules,
  region_DA_or_DB_IDs,
  themes = cc.data::list_accessibility_themes(),
  traveltimes,
  time_intervals = which(1:60\%\%5 == 0),
  pe_include = c("access_foot_educational_elementary",
    "access_bicycle_educational_secondary", "access_foot_cultural_total",
    "access_foot_food_grocery", "access_car_healthcare_hospitals"),
  default_var = "access_foot_food_grocery",
  modes = c("foot", "bicycle", "car", "transit"),
  scales_sequences,
  crs,
  DA_DB = "DA",
  overwrite = FALSE,
  inst_prefix
)
}
\arguments{
\item{scales_variables_modules}{<\verb{named list}> A list of length three.
The first is all the scales, the second is the variables table, and the
third is the modules table.}

\item{region_DA_or_DB_IDs}{<\verb{character vector}> All the current census'
DA/DB IDs present in the region. Only those will be extracted from the database.}

\item{themes}{<\verb{character vector}> All pre-processed themes to import.
Defaults to everything: \code{cc.data::accessibility_themes}}

\item{traveltimes}{<\verb{named list}>A list of matrices containing travel times
between origin-destination pairs. The output of
\code{\link[cc.buildr]{accessibility_get_travel_times}}}

\item{time_intervals}{<\verb{numeric vector}> A vector of time intervals
(in minutes) to calculate travel times for. Defaults to \code{c(5, 10, 15, ..., 60)}. 60 minutes is the maximum.}

\item{pe_include}{<\verb{character vector}> Which final variables should appear
in the place explorer. Defaults to \code{c("access_foot_20_educational_elementary", "access_foot_20_communitycentres_individual", "access_foot_15_fooddistribution_grocery", "access_car_10_healthcare_hospitals")}.}

\item{default_var}{<\code{character}> The first variable the user will see when
they will lang on the page. Defaults to groceries accessible within a 20 minutes walk.}

\item{modes}{<\verb{character vector}> For which mode should accessibility be calculated?
Defaults to all: \code{c("foot", "bicycle", "car", "transit")}}

\item{scales_sequences}{<\code{list}> A list of scales sequences representing the
hierarchical ordering of scales on an auto-zoom.}

\item{crs}{<\code{numeric}> EPSG coordinate reference system to be assigned, e.g.
\code{32617} for Toronto.}

\item{DA_DB}{<\code{character}> Which of DA or DB should be used to calculate accessibility.}

\item{overwrite}{<\code{logical}> Should the data already processed and stored be
overwriten?}

\item{inst_prefix}{<\code{character}> The prefix of the instance, e.g. \code{'mtl'} which
is the database schema in which the data is saved.}
}
\value{
A list containing the scales, variables, and modules tables.
}
\description{
Calculates the average count of amenities (such as schools, food distributors,
health care facilities, etc.) accessible by walk, bike, transit, or car within a
given amount of time and add it to \code{scales_variables_modules}
}
