% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_age.R
\name{ba_age}
\alias{ba_age}
\title{Build and append age from census data}
\usage{
ba_age(
  scales_variables_modules,
  scales_sequences,
  scales_to_interpolate,
  overwrite = FALSE
)
}
\arguments{
\item{scales_variables_modules}{<\verb{named list}> A list of length three.
The first is all the scales, the second is the variables table, and the
third is the modules table.}

\item{scales_sequences}{<\code{list}> A list of scales sequences representing the
hierarchical ordering of scales on an auto-zoom.}

\item{scales_to_interpolate}{<\verb{character vector}> WILL NOT be used for calculations.
Simply used to inform the variables table. ALL scales holding census \code{age} data
will be used for calculation of the age page data.}

\item{overwrite}{<\code{logical}> Should the data already precessed and stored be
overwriten?}
}
\value{
A list of length 4, similar to the one fed to
\code{scales_variables_modules} with age data added, their addition
in the variables table and the module table.
}
\description{
Build and append age from census data
}
