% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_and_append_census_data.R
\name{ba_census_data}
\alias{ba_census_data}
\title{Build and append census data}
\usage{
ba_census_data(scales_variables_modules, crs, housing_module = TRUE)
}
\arguments{
\item{scales_variables_modules}{<\verb{names list}> A list of length three.
The first is all the scales, the second is the variables table, and the
third is the modules table.}

\item{crs}{<\code{numeric}> EPSG coordinate reference system to be assigned, e.g.
\code{32617} for Toronto.}

\item{housing_module}{<\code{logical}> Should a housing module be added to
the list of modules.}
}
\description{
Build and append census data
}
