% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_ndvi.R
\name{ba_ndvi}
\alias{ba_ndvi}
\title{Build and append NDVI (Normalized Difference Vegetation Index) data}
\usage{
ba_ndvi(
  scales_variables_modules,
  data_output_path = "dev/data/ndvi/",
  years = cc.data::ndvi_years(),
  skip_scales = NULL,
  scales_sequences,
  crs
)
}
\arguments{
\item{scales_variables_modules}{<\verb{named list}> A list of length three.
The first is all the scales, the second is the variables table, and the
third is the modules table.}

\item{data_output_path}{<\code{character}> String representing the directory path
to store the NDVI data (default is "dev/data/ndvi/").}

\item{years}{<\verb{numeric vector}> Years for which to build the data. Defaults to
\code{cc.data::ndvi_years()}.}

\item{skip_scales}{<\verb{character vector}> Scales to be skipped in the analysis.
Already skipped scales are building and street. Defaults to NULL.}

\item{scales_sequences}{<\code{list}> A list of scales sequences representing the
hierarchical ordering of scales on an auto-zoom.}

\item{crs}{<\code{numeric}> EPSG coordinate reference system to be assigned, e.g.
\code{32617} for Toronto.}
}
\value{
A list of length 3, similar to the one fed to
\code{scales_variables_modules} with the NDVI variable added, its addition
in the variables table and the module table.
}
\description{
This function downloads and import Harmonized Landsat Sentinel-2 raster data
for the master_polygon at 30x30m. The NDVI is calculated from it and gets
averaged for every features in \code{all_scales}. It includes data retrieval, transformations,
filtering, merging, and calculation of region values and breaks.
}
