% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BC_page_bill44.R
\name{bill44_page}
\alias{bill44_page}
\title{Bill 44 Zoning Analysis Page Function}
\usage{
bill44_page(
  scales_variables_modules,
  scales_sequences,
  crs,
  overwrite,
  inst_prefix
)
}
\arguments{
\item{scales_variables_modules}{<\code{list}> A list containing scales, variables, and
modules data structures.}

\item{scales_sequences}{<\code{list}> A list of scale sequences for data representation.}

\item{crs}{<\code{numeric}> Coordinate reference system to be used for spatial data.}

\item{overwrite}{<\code{logical}> Flag to determine if existing data should be overwritten.}

\item{inst_prefix}{<\code{character}> The prefix of the instance, e.g. \code{'mtl'} which
is the database schema in which the data is saved.}
}
\value{
<\code{list}> A list containing updated scales, variables, and modules.
}
\description{
This function performs a zoning analysis based on Bill 44 legislation, focusing
on the impact of small-scale, multi-unit housing in various zones. It integrates
geographic data transformation, intersection analysis, and potential housing unit
calculations. The function finally updates modules and variables tables.
}
