% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_buildings.R
\name{build_buildings}
\alias{build_buildings}
\title{Build buildings dataframe}
\usage{
build_buildings(DA_table, crs, download_MS_buildings = TRUE, MS_province)
}
\arguments{
\item{DA_table}{<\code{data.frame}> A \code{DA} data.frame from which boundaries
are used to filter the buildings, and add other DA information.}

\item{crs}{<\code{numeric}> EPSG coordinate reference system to be assigned, e.g.
\code{32618} for Montreal.}

\item{download_MS_buildings}{<\code{logical}> Should the Microsoft's Canadian
Building Footprints dataset be used to add the missing buildings to the OSM data?}

\item{MS_province}{<\code{character}> The Province for which Microsoft's buildings
should be download. One of: \code{"Alberta"}, \code{"BritishColumbia"},
\code{"Manitoba"}, \code{"NewBrunswick"}, \code{"NewfoundlandAndLabrador"},
\code{"NorthwestTerritories"}, \code{"NovaScotia"}, \code{"Nunavut"},
\code{"Ontario"}, \code{"PrinceEdwardIsland"}, \code{"Quebec"},
\code{"Saskatchewan"}, \code{"YukonTerritory"}.}
}
\value{
Returns a data.frame of all buildings in the boundaries of the
\code{"DA_table"}, with an \code{"ID"}, \code{"name_2"}, \code{"DAUID"},
\code{"CTUID"}, \code{"CSDUID"}, \code{"osm_ID"} and, \code{"geometry"} column.
}
\description{
\code{build_buildings()} uses the a DA table to download building data from
Open Street Map and from Microsoft's Canadian Building Footprints.
}
