% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_census_data.R
\name{build_census_data}
\alias{build_census_data}
\title{Build census data at all possible scales}
\usage{
build_census_data(scales_consolidated, census_data, census_data_raw_DA, crs)
}
\arguments{
\item{scales_consolidated}{A named list of sf data.frame
containing all scales listed with their regions. The output of
\code{\link[susbuildr]{consolidate_scales}}.}

\item{census_data}{"TKTK"}

\item{census_data_raw_DA}{"TKTK"}

\item{crs}{<\code{numeric}> EPSG coordinate reference system to be assigned, e.g.
\code{32617} for Toronto.}
}
\value{
Returns a list of length 3. The first is the possible scales for which
census data can be added in scales_consolidated. The second is a
data.frame of scales reference, to know for which scales the data is
available. It can directly be fed to the scales argument of add_variable.
The third is a data.frame of interpolation reference, to know for which
scales the data has been interpolated. It can directly be fed to the
interpolated argument of add_variable.
}
\description{
Build census data at all possible scales
}
