% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_census_scales.R
\name{build_census_scales}
\alias{build_census_scales}
\title{Build census scales}
\usage{
build_census_scales(
  master_polygon,
  census_dataset,
  regions,
  levels = c("CSD", "CT", "DA"),
  crs,
  fill_CTs_with_CSDs = TRUE,
  split_CSD,
  default_name_2_for_CSD = "City"
)
}
\arguments{
\item{master_polygon}{<\code{sfc_MULTIPOLYGON}>. Unioned multipolygon of all the
geometries for which census data must be gathered.}

\item{census_dataset}{<\code{character}> The dataset to query for available
regions, e.g. \code{"CA16"}.}

\item{regions}{<\verb{named list}> A named list of census regions to retrieve.
Names must be valid census aggregation levels. Preferably a whole province
to make sure all geometries present in the <\code{master_polygon}> is added.
e.g. \code{list(PR = 24)} for Montreal.}

\item{levels}{<\code{character}> The census aggregation levels to retrieve.
Can be \code{"CMA"}, \code{"CD"}, \code{"CT"}, \code{"DA"}, \code{"DB"}, ...}

\item{crs}{<\code{numeric}> EPSG coordinate reference system to be assigned, e.g.
\code{32618} for Montreal.}

\item{fill_CTs_with_CSDs}{<\code{logical}> If the master polygon is bigger than a
CMA, should what is outside the CMA display the CSDs on the CT scale?}

\item{split_CSD}{<\verb{sf data.frame}> An sf dataframe of subdivisions of one
CSD. e.g. for the City of Montreal, the subdivisions is a dataframe of boroughs.
The sf data.frame must have 3 columns: \code{name}, \code{type} and \code{geometry}.
The \code{name} is the name of the borough/neighbourhood/zone e.g.
"Le Plateau Mont-Royal", and \code{type} is what should be used as \code{"name_2"},
e.g. "Borough". The data.frame must fit with CTs boundaries.}

\item{default_name_2_for_CSD}{<\code{character}> The default name_2 for CSDs. Default
to "City". The display on Sus would then be e.g. \verb{City of Laval}.}
}
\value{
A list of sf dataframes of census scales filtered by the master polygon,
with the option of one CSD subdivided.
}
\description{
\code{build_census_scales} retrieves all the fed census levels
present in the master_polygon. It can split a CSD in multiple smaller zones.
The \code{split_CSD} fed in the function then sums population and households for
each of its zones from the census tracts, and the geometry of the zones are
swaped by an aggregation of census tracts for which centroids fall in each zones,
so that borders of the split zones fit perfectly with the rest of the CSD.
Boundaries of both the zones and the overall split_CSD dataframe MUST fit
with census tract boundaries.
}
