% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{build_colours}
\alias{build_colours}
\title{Build color palettes for data visualization}
\usage{
build_colours(
  left_5 = c("#C4CDE1", "#98A8CB", "#6C83B5", "#4C5C7F", "#2B3448"),
  left_3 = c("#E8E8E8", "#B5C0DA", "#6C83B5"),
  right_5 = c("#C7DFCC", "#9DC6A6", "#73AE80", "#517A5A", "#2E4633"),
  right_3 = c("#E8E8E8", "#B8D6BE", "#73AE80"),
  delta_5 = c("#6D5271", "#DCD3E2", "#E8E8E8", "#f9e699", "#F69036"),
  delta_neg_5 = c((grDevices::colorRampPalette(delta_5[c(1, 2)]))(4), "#E8E8E8"),
  delta_pos_5 = c("#E8E8E8", (grDevices::colorRampPalette(delta_5[c(4, 5)]))(4)),
  bivar = c("#E8E8E8", "#B5C0DA", "#6C83B5", "#B8D6BE", "#90B2B3", "#567994", "#73AE80",
    "#5A9178", "#2A5A5B"),
  qual = c("#73AE80", "#6C83B5", "#5B362A", "#B58A6C", "#2A5A5B", "#AE7673"),
  col_NA = c("#B3B3BB"),
  viridis = (scales::viridis_pal())(25)
)
}
\arguments{
\item{left_5}{<\verb{colours vector}> for a sequential scale of 5 values
for the left side of the color scale.}

\item{left_3}{<\verb{colours vector}> for a sequential scale of 3 values
for the left side of the color scale.}

\item{right_5}{<\verb{colours vector}> for a sequential scale of 5 values
for the right side of the color scale.}

\item{right_3}{<\verb{colours vector}> for a sequential scale of 3 values
for the right side of the color scale.}

\item{delta_5}{<\verb{colours vector}> scale used for variation of 5 values.
By defaults starts with red and go to blue.}

\item{bivar}{<\verb{colours vector}> for a bivariate scale of 9 values, by default
the stevens.greenblue palette.}

\item{qual}{<\verb{colours vector}> for a qualitative scale of 6 values.}

\item{col_NA}{<\verb{colours vector}> for missing values.}

\item{viridis}{<\verb{colours vector}> for a viridis color scale.}
}
\value{
A list of tibbles with color scales to be used for the rdeck map,
legend and other plots.
}
\description{
This function generates a set of color palettes for data visualization
purposes. It returns a list containing six data frames that can be used
for setting color scales in rdeck, ggplot or other visualization packages.
}
