% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_postal_codes.R
\name{build_postal_codes}
\alias{build_postal_codes}
\title{Retrieve the postal codes in the region}
\usage{
build_postal_codes(region_DA_IDs)
}
\arguments{
\item{region_DA_IDs}{<\verb{character vector}> All the current census'
DA IDs present in the region. Only those will be extracted from the database
to do interpolation.}
}
\value{
Returns all the postal codes in the region along with their centroid.
}
\description{
Retrieve the postal codes in the region
}
