% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_streets.R
\name{build_streets}
\alias{build_streets}
\title{Build streets dataframe}
\usage{
build_streets(DA_table, crs)
}
\arguments{
\item{DA_table}{<\verb{sf data.frame}> A \code{DA} sf data.frame from which boundaries
are used to filter the streets, and add other DA information.}

\item{crs}{<\code{numeric}> EPSG coordinate reference system to be assigned, e.g.
\code{32618} for Montreal.}
}
\value{
Returns a data.frame of all street segments in the boundaries of the
\code{"DA_table"}, with \code{"ID"}, \code{"name_2"}, \code{"DAUID"},
\code{"osm_ID"} and, \code{"geometry"} columns.
}
\description{
\code{build_streets()} uses a DA table to download street data from
Open Street Map.
}
