% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{calculate_breaks}
\alias{calculate_breaks}
\title{Calculate all breaks when all_scales is a list of regions and scales}
\usage{
calculate_breaks(
  all_scales,
  vars,
  time_regex = "_\\\\d{4}$",
  types = NULL,
  rank_name = c("Very low", "Low", "Moderate", "High", "Very high"),
  rank_name_short = c("Very low", "Low", "Moderate", "High", "Very high"),
  breaks_base = NULL,
  use_quintiles = FALSE
)
}
\arguments{
\item{all_scales}{<\code{named_list}> A named list of scales. The first level is
the geo, and the second is the scales. They must contain all columns of \code{vars}}

\item{vars}{<\verb{vector of character}> Contains all variable names from which
to add q3s. Must fit with a name in \code{df}.}

\item{time_regex}{<\code{character}> Regular expression which corresponds to
a timeframe, placed at the end of the \code{vars} vector. e.g. \verb{_\\\\d\{4\}$} for
years. If the variable does not have a timeframe, enter an empty string (\code{""}).}

\item{types}{<\code{list}> A named list of variable types (e.g., "pct", "avg", "count", "ind").
The names of the list should match the variable names in \code{vars} (without the dates).}

\item{rank_name}{<\verb{character vector}> If the type is \code{ind}, how should every ranks
from 1 to 5 (q5) be named. Defaults to c("Low", "Below average", "Average",
"Above average", "High")}

\item{rank_name_short}{<\verb{character vector}> Same as rank_name but shorter
to not take too much space in the legend labels. Defaults to c("Low", "B. average",
"Average", "A. average", "High")}

\item{breaks_base}{<\verb{names list}> If we want variables to use the breaks of
another variable, we can supply a named list. The name is the variable that will
use the breaks of the value variable. ex. in access, we want all variables of
same theme to use the same breaks, no matter the transportation time. This
would be how we would do the list, if we want to use the 30 minutes break as
the basis for all the other transport time.
\code{list(access_transit_pwd_25_food_misc_2023 = "access_transit_pwd_30_food_misc_2023", access_transit_pwd_20_food_misc_2023 = "access_transit_pwd_30_food_misc_2023", ...)}}

\item{use_quintiles}{<\code{logical}> Should the breaks be pretty and constant made
by splitting the data in 5 using the minimum and maximal value, or should we use
a quintile approach. The latter is better suited for non-normal distributions.}
}
\value{
Returns a list of length 4. The first is the same data.frame as df
with q3 and q5 columns appended. The second is the q3 breaks table, and the third
is the q5 breaks table. The fourth is vectors of characters of all dates at
which variables are available for.
}
\description{
Calculate all breaks when all_scales is a list of regions and scales
}
