% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{canale_data}
\alias{canale_data}
\title{2016 Dissemination Areas Can-ALE Index}
\format{
\subsection{\code{centraide_data}}{

A data frame of 2 columns and the same number of rows as there are DAs in
the 2016 Canadian Census.
\describe{
\item{DA_ID}{The ID of each dissemination area}
\item{canale_2016}{The 2016 Can-ALE index}
}
}
}
\source{
\url{http://canue.ca/wp-content/uploads/2018/03/CanALE_UserGuide.pdf}
}
\usage{
canale_data
}
\description{
Canadian Active Living Environments Database.
}
\keyword{datasets}
