% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{canbics_data}
\alias{canbics_data}
\title{2021 Dissemination Areas Can-BICS Metric}
\format{
\subsection{\code{canbics_data}}{

A data frame of 2 columns and the same number of rows as there are DAs in
the 2016 Canadian Census.
\describe{
\item{DA_ID}{The ID of each dissemination area}
\item{canbics_2021}{The 2021 Can-BICS metric}
}
}
}
\source{
\url{https://canue.ca/wp-content/uploads/2022/04/CAN-BICS_UserGuide_format.docx}
}
\usage{
canbics_data
}
\description{
Neighborhood Canadian Bikeway Comfort and Safety Classification System
Database.
}
\keyword{datasets}
