% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_scales_dictionary.R
\name{census_scales_dictionary}
\alias{census_scales_dictionary}
\title{Build census scales dictionary}
\usage{
census_scales_dictionary(census_scales)
}
\arguments{
\item{census_scales}{<\code{list}> A named list containing census scales built
with the \code{\link[susbuildr]{build_census_scales}} function.}
}
\value{
A data.frame used as dictionary for texts on the platform.
}
\description{
Build census scales dictionary
}
