% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_scales.R
\name{consolidate_scales}
\alias{consolidate_scales}
\title{Consolidate Scales}
\usage{
consolidate_scales(scales_sequences, all_scales, regions, crs)
}
\arguments{
\item{scales_sequences}{<\code{list}> A list of scales sequences representing the
hierarchical ordering of scales on an auto-zoom.}

\item{all_scales}{<\verb{named list}> A list of data frames corresponding to each
scale.}

\item{regions}{<\verb{named list}> A list or unioned polygons of regions for which the
consolidated data needs to be processed.}

\item{crs}{<\code{character}> A Coordinate Reference System.}
}
\value{
A list containing:
\itemize{
\item \code{scales}: A modified list of all scales with corrected identifiers.
\item \code{for_region_dict_scales}: A list of scale IDs for each region,
spatially filtered based on 1\% threshold.
}
}
\description{
This function takes all the scales and ensure uniformity. It also notes,
using the sequences of scales, all the IDs of scales above them. It also
calculates all the features which are part of regions.
}
