% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_svg_to_ico.R
\name{convert_svg_to_ico}
\alias{convert_svg_to_ico}
\title{Convert SVG to ICO}
\usage{
convert_svg_to_ico(
  new_color,
  svg_path = paste0("https://s3.amazonaws.com/curbcut.pu",
    "blic.resources/Curbcut-symbol_c-imp", "act-black.svg"),
  ico_path = "www/favicon.ico",
  ico_size = 32,
  save_logos = TRUE
)
}
\arguments{
\item{new_color}{<\code{character}> Hexadecimal code for the new color.}

\item{svg_path}{<\code{character}> URL or path to the SVG file. Default is a
specific URL pointing to a Curbcut symbol.}

\item{ico_path}{<\code{character}> Path where the ICO file will be saved. Default
is "www/favicon.ico".}

\item{ico_size}{<\code{numeric}> Size for the ICO file in pixels. Default is 32.}

\item{save_logos}{<\code{logical}> Whether to save additional logo images at 192x192
and 512x512 pixels. Default is TRUE.}
}
\value{
The function does not return a value but saves the ICO file and
optionally the logo images at the specified paths.
}
\description{
This function converts an SVG image to an ICO format. It also allows for
changing the color of the SVG and resizing it to a specified size. The
function can optionally save logo images in different sizes.
}
