% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_script.R
\name{create_data_script}
\alias{create_data_script}
\title{Create or edit a build and append data script}
\usage{
create_data_script(name, pages_folder = "dev/data_import/")
}
\arguments{
\item{name}{<\code{character}> Name of the data script to be created, e.g. \code{canale}.}

\item{pages_folder}{<\code{character}> Path to the folder where all the build and
append data scripts are in.}
}
\value{
Opens a build and append data script.
}
\description{
Create or edit a build and append data script
}
