% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home_page.R
\name{create_home_page_script}
\alias{create_home_page_script}
\title{Create a home page script (buidl and save necessary tibbles)}
\usage{
create_home_page_script(pages_folder = "dev/data_import/")
}
\arguments{
\item{pages_folder}{<\code{character}> Path to the folder where all the build and
append data scripts are in.}
}
\value{
Opens a home page data building script.
}
\description{
Create a home page script (buidl and save necessary tibbles)
}
