% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_story.R
\name{create_story_rmd}
\alias{create_story_rmd}
\title{Create an Rmd file for a story}
\usage{
create_story_rmd(name, stories_folder = "dev/Rmd/stories/")
}
\arguments{
\item{name}{The name of the story}

\item{stories_folder}{The folder where the story should be created (
defaults to "dev/Rmd/stories/")}
}
\value{
Opens an empty new story .Rmd (pre-populated with a template)
}
\description{
Create an Rmd file for a story
}
