% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_story.R
\name{create_story_script}
\alias{create_story_script}
\title{Create a story script}
\usage{
create_story_script(pages_folder = "dev/data_import/")
}
\arguments{
\item{pages_folder}{The folder where the story script will be created.
Defaults to \code{"dev/data_import/"}}
}
\value{
Open a story script in the given folder
}
\description{
Create a story script
}
