% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{current_census}
\alias{current_census}
\title{Current census code, used to retrieve from cancensus}
\format{
\subsection{\code{current_census}}{

A character vector of the most up-to-date available census from cancensus.
}
}
\usage{
current_census
}
\description{
Current census code, used to retrieve from cancensus
}
\keyword{datasets}
