% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{da_centroids_popw}
\alias{da_centroids_popw}
\title{2016 Population-weighted Dissemination Area Centroids}
\format{
\subsection{\code{da_centroids_popw}}{

An sf data frame with 7,240 56590 and 2 columns:
\describe{
\item{ID}{The ID of each dissemination area}
\item{geometry}{sfc_POINT representing the population-weighted centroid
of each dissemination area}
}
}
}
\usage{
da_centroids_popw
}
\description{
The centroids of all dissemination areas, weighted by population, in the
country.
}
\keyword{datasets}
