% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_construct.R
\name{data_construct}
\alias{data_construct}
\title{Append new data to the \code{data} element of scales_variables_modules}
\usage{
data_construct(
  svm_data,
  scales_data,
  unique_var,
  time_regex,
  schema = list(time = time_regex),
  breaks_var = NULL
)
}
\arguments{
\item{svm_data}{<\verb{named list}> Lits of already existing data (usually
\code{scales_variables_modules$data})}

\item{scales_data}{<\verb{named list}> A list containing the new data for each scale.}

\item{unique_var}{<\code{character}> A character vector of unique variable names to
construct data for, e.g. \code{c('alp', 'housing_tenant', ...)} without the
time.}

\item{time_regex}{<\code{character}> Regular expression which corresponds to
a timeframe, placed at the end of the \code{vars} vector. e.g. \verb{_\\\\d\{4\}} for
years.}

\item{schema}{<\verb{named list}> Named list of schemas available for the live app.
It must be a regular expression grabbing said info from the variable name. e.g.
\code{list(time = '\\\\d{4}$')} for grabbing year.}

\item{breaks_var}{<\verb{named list}> The specific variable from which to draw the
breaks out of all the values. Every element of the list is using a name
of \code{unique_var} and has as value a complete variable code with schemas. e.g.
for unique_var \code{alp}, breaks_var would be \code{list(alp = "alp_2021")}. Defaults to
NULL which will default in using the last ordered column (latest date).}
}
\value{
A merged list containing the processed data tables for each scale and
variable.
}
\description{
This function processes the provided scales data, constructs a new data table
based on the specified unique variables, and integrates additional time-related
attributes. It further tests for binning breaks and whether quintiles should
be used. The resulting data is merged with the existing scales_variables_modules data.
}
\details{
If bin occurrences suggest uneven distributions (one bin with more than 50\%
of observations), quintiles are applied. If breaks calculations
are not adequate, the function stops with an error.
}
