% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_postgresql.R
\name{db_connect_prod}
\alias{db_connect_prod}
\title{Connect to the AWS PostgreSQL database}
\usage{
db_connect_prod()
}
\value{
returns an S4 object that inherits from \code{\link[DBI]{DBIConnection-class}}.
This object is used to communicate with the database engine.
}
\description{
Connect to the AWS PostgreSQL database using variables saved in the .Renviron
}
