% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_postgresql.R
\name{db_create_schema_prod}
\alias{db_create_schema_prod}
\title{Create or Verify existence of a PostgreSQL schema}
\usage{
db_create_schema_prod(schema_name, allow_read_webapp = TRUE)
}
\arguments{
\item{schema_name}{<\code{character}> The name of the schema to check or create.
Expected to be a character string representing the valid schema name in PostgreSQL.}

\item{allow_read_webapp}{<\code{logical}> Wether if the read only webapp user in the
database should have read access to this schema. Defaults to TRUE.}
}
\value{
A character string message indicating whether the schema was created or
already existed.
}
\description{
Connects to the PostgreSQL database, checks if the specified schema exists,
and creates it if it does not. If the schema already exists, it does nothing except
inform the user. It uses a database connection, defined by \code{db_connect_prod()},
and ensures safe execution and disconnection using \code{db_try_disconnect()}.
Normally, as schema is used for an instance of Curbcut. mtl, to, van, etc.
}
