% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_postgresql.R
\name{db_disconnect_prod}
\alias{db_disconnect_prod}
\title{Disconnect from the AWS PostgreSQL database}
\usage{
db_disconnect_prod(conn)
}
\arguments{
\item{conn}{A \code{\link[DBI]{DBIConnection-class}} object, as returned by
\code{\link[cc.data]{db_connect}}.}
}
\value{
A message out of \code{\link[DBI]{dbDisconnect}}
}
\description{
Disconnect from the AWS PostgreSQL database
}
