% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_postgresql.R
\name{db_list_scales}
\alias{db_list_scales}
\title{List scales in a PostgreSQL schema}
\usage{
db_list_scales(schema)
}
\arguments{
\item{schema}{<\code{character}> name of the schema whose tables you want to list.}
}
\value{
A character vector containing the names of all tables (scales) within the
specified schema.
}
\description{
Retrieves a list of all tables within a specified schema in a PostgreSQL
database. The PostgreSQL database is accessed through \code{db_connect_prod}, and only
tables which DO NOT contain underscores are extracted (they are Curbcut \code{scales}).
}
