% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_postgresql.R
\name{db_list_tables}
\alias{db_list_tables}
\title{List Tables in a PostgreSQL Schema}
\usage{
db_list_tables(schema)
}
\arguments{
\item{schema}{<\code{character}> name of the schema whose tables you want to list.}
}
\value{
A character vector containing the names of all tables within the
specified schema.
}
\description{
Retrieves a list of all tables within a specified schema in a PostgreSQL
database. The PostgreSQL database is accessed through \code{db_connect_prod}.
}
