% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_postgresql.R
\name{db_list_tables_of_scale}
\alias{db_list_tables_of_scale}
\title{List Tables in a PostgreSQL Schema Starting with 'grd30'}
\usage{
db_list_tables_of_scale(schema, scale)
}
\arguments{
\item{schema}{<\code{character}> name of the schema whose tables you want to list.}

\item{scale}{<\code{character}> name of the scale for which we want to list all
the data tables}
}
\value{
A character vector containing the names of tables within the
specified schema that start with 'grd30'.
}
\description{
Retrieves a list of all tables within a specified schema in a PostgreSQL
database that start with 'grd30'. The PostgreSQL database is accessed
through \code{db_connect_prod}.
}
