% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_postgresql.R
\name{db_try_disconnect}
\alias{db_try_disconnect}
\title{Try to execute a database function and disconnect on error}
\usage{
db_try_disconnect(conn, fun)
}
\arguments{
\item{conn}{A database connection object. This should be an object
representing an open connection to a database, as created by
\code{DBI::dbConnect()}.}

\item{fun}{A function that performs an operation on the database using
the provided \code{conn}. This function should use the connection to
execute a query, fetch results, write data, etc.}
}
\value{
The result of executing \code{fun}, if successful. If an error
occurs during the execution of \code{fun}, the database connection is
closed, and the error is re-thrown.
}
\description{
This function attempts to execute a given database-related function
(\code{fun}) using an existing database connection (\code{conn}). If an error
occurs during the execution of \code{fun}, the function ensures that the
database connection is properly disconnected by calling
\code{db_disconnect_prod}. This mechanism is useful for maintaining clean
database connections and ensuring they are closed properly in the event
of an error.
}
