% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_postgresql.R
\name{db_write_prod}
\alias{db_write_prod}
\title{Write dataframe to PostgreSQL with Primary Key}
\usage{
db_write_prod(
  df,
  table_name,
  schema,
  scale = NULL,
  primary_key = "ID",
  index = NULL
)
}
\arguments{
\item{df}{<\code{data.frame}> A data frame containing the data to be written to the
database table.}

\item{table_name}{<\code{character}> A string specifying the name of the database table.}

\item{schema}{<\code{character}> A string specifying the database schema, which is
the local prefix which is also the same prefix of tilesets. E.g. for Montreal, \code{mtl}.}

\item{scale}{<\code{character}> The scale at which is the data. It is coupled
with \code{table_name} for the name of the table in the database. If \code{NULL} (default),
only \code{table_name} is used. It stays NULL only for the general scale table, not
for table containing data.}

\item{primary_key}{<\code{character}> A string specifying the name of the column
to be used as the primary key. Defaults to "ID".}

\item{index}{<\code{character}> or <\verb{character vector}> specifying the name(s) of
the column(s) to be indexed. If \code{NULL} (default), no additional indexes are created.
Multiple columns can be provided for creating a composite index. Defaults to \code{NULL}.}
}
\value{
Invisibly returns \code{NULL}. The primary effect is the side effect of
writing data to the database and setting a primary key.
}
\description{
This function writes a given data frame to a specified table in a PostgreSQL
database. If the table exists, it is overwritten. The function also sets a
specified column as the primary key of the table. The operation is wrapped in
\code{db_try_disconnect} to ensure the database connection is properly managed.
}
