% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heroku.R
\name{docker_add_env}
\alias{docker_add_env}
\title{Add Environment Variables to a Dockerfile}
\usage{
docker_add_env(dockerfile_path = "Dockerfile")
}
\arguments{
\item{dockerfile_path}{<\code{character}> Path to the Dockerfile. Default
is "Dockerfile". The function updates this Dockerfile by inserting
environment variable definitions.}
}
\value{
<\code{logical}> Returns nothing. In case of missing environment variables,
the function halts with an error message.
}
\description{
This function adds specified environment variables to a Dockerfile,
essential for configuring the container's environment. It targets a
specific insertion point before the installation of R environment
(renv) packages. Environment variables related to both the project's
database and AWS credentials are set, ensuring necessary services
are correctly configured within the Docker container.
}
