% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyk.R
\name{dyk_prep}
\alias{dyk_prep}
\title{Prepare a Variable Table for DYK Generation}
\usage{
dyk_prep(svm, all_tables, n = NULL)
}
\arguments{
\item{svm}{<\code{list}> A list, usually \code{scales_variables_modules}, containing
the scales, modules, and variables tables.}

\item{all_tables}{<\code{list}> A list, usually \code{all_tables}, containing the
scales present in each region.}

\item{n}{<\code{integer}> Optionally, an integer scalar specifying the number of
rows in \code{svm$modules} to be used for creating the DYK variable table. The
default (\code{NULL}) will use all rows; other values may be useful for testing.}
}
\value{
A data frame with six columns (\code{module}, \code{region}, \code{scale}, \code{date},
\code{var_left} and \code{var_right}), each of which is a character vector.
}
\description{
This function creates a table of variable combinations which serves as the
input to the "Did you know" generation process.
}
