% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyk.R
\name{dyk_uni}
\alias{dyk_uni}
\title{Create a Table of Univariate DYKs}
\usage{
dyk_uni(vars_dyk, svm, scales_dictionary, langs, translation_df = NULL)
}
\arguments{
\item{vars_dyk}{<\code{data.frame}> A data frame with columns \code{module},
\code{region}, \code{scale}, \code{var_left}, \code{var_right} and \code{date}, probably created with
\code{\link[cc.buildr]{dyk_prep}}.}

\item{svm}{<\code{list}> A list, usually \code{scales_variables_modules}, containing
the scales, modules, and variables tables.}

\item{scales_dictionary}{<\code{data.frame}> A data frame with information about
scale names and text strings}

\item{langs}{<\code{character}> A character vector specifying for which languages
DYK strings should be produced. Allowable values are "en" and "fr".}

\item{translation_df}{<\code{data.frame}> A data frame with columns for each
language present in the \code{langs} argument.}
}
\value{
A data frame with ten columns (the columns present in \code{vars_dyk},
and then additionally \code{select_ID}, \code{dyk_type}, \code{dyk_text} and \code{dyk_weight}).
}
\description{
This function creates a table of univariate "Did you know" text strings.
}
