% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyk.R
\name{dyk_uni_compare}
\alias{dyk_uni_compare}
\title{Generate Bivariate Comparison DYKs}
\usage{
dyk_uni_compare(var_left, var_right, region, scale, date, svm)
}
\arguments{
\item{var_left}{\if{html}{\out{<character>}} A string representing the name of the first
variable for which the DYK should be calculated.}

\item{var_right}{\if{html}{\out{<character>}} A string representing the name of the second
variable for which the DYK should be calculated.}

\item{region}{<\code{character}> A string representing the name of the region
for which the DYK should be calculated.}

\item{scale}{<\code{character}> A string representing the name of the scale
for which the DYK should be calculated.}

\item{date}{<\code{character}> A string representing the name of the date
for which the DYK should be calculated.}

\item{svm}{<\code{list}> A list, usually \code{scales_variables_modules}, containing
the scales, modules, and variables tables.}
}
\value{
A data frame with two columns (\code{compare_text} and \code{compare_val}),
which contain a character vector of DYK outputs and a numeric vector of the
values respectively.
}
\description{
This function creates "Did you know" text strings highlighting bivariate
comparisons from a combination of variables, region, scale and date.
The output is a data frame containing a \code{highest} and \code{lowest} column.
}
