% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyk.R
\name{dyk_uni_highest_lowest}
\alias{dyk_uni_highest_lowest}
\title{Generate Highest/Lowest Value DYKs}
\usage{
dyk_uni_highest_lowest(
  var_left,
  region,
  scale,
  date,
  svm,
  scales_dictionary,
  regions_dictionary,
  langs
)
}
\arguments{
\item{var_left}{\if{html}{\out{<character>}} A string representing the name of the variable
for which the DYK should be calculated.}

\item{region}{<\code{character}> A string representing the name of the region
for which the DYK should be calculated.}

\item{scale}{<\code{character}> A string representing the name of the scale
for which the DYK should be calculated.}

\item{date}{<\code{character}> A string representing the name of the date
for which the DYK should be calculated.}

\item{svm}{<\code{list}> A list, usually \code{scales_variables_modules}, containing
the scales, modules, and variables tables.}

\item{scales_dictionary}{<\code{data.frame}> A data frame with information about
scale names and text strings}

\item{regions_dictionary}{<\code{data.frame}> A data frame with information about
which feature of every scale fits in which region.}

\item{langs}{<\code{character}> A character vector specifying for which languages
DYK strings should be produced. Allowable values are "en" and "fr".}
}
\value{
A data frame with four columns (\code{highest}, \code{lowest}, \code{highest_ID},
and \code{lowest_ID}), each of which contains a character vector of DYK outputs.
}
\description{
This function creates "Did you know" text strings highlighting the highest
and lowest values from a combination of variable, region, scale and date.
The output is a data frame containing a \code{highest} and \code{lowest} column.
}
