% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data_files.R
\name{exclude_processed_scales}
\alias{exclude_processed_scales}
\title{Exclude scales with processed data already saved}
\usage{
exclude_processed_scales(
  unique_vars,
  scales,
  overwrite = FALSE,
  data_folder = "data/"
)
}
\arguments{
\item{unique_vars}{<\verb{character vector}> A vector of unique variable names.}

\item{scales}{<\verb{character vector OR named list}> A vector of scale names or a named
list of scales to be checked.}

\item{overwrite}{<\code{logical}> If TRUE, no scales are excluded and all are returned.}

\item{data_folder}{<\code{character}> The folder where data files are stored.
Default is "data/".}
}
\value{
<\verb{character vector}> Scales for which data files do not exist
or all scales if 'overwrite' is TRUE.
}
\description{
This function excludes scales for which data has already been processed and stored.
If 'overwrite' is TRUE, no exclusion is performed and all scales are returned.
}
