% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{find_breaks_q5}
\alias{find_breaks_q5}
\title{Find pretty q5 breaks}
\usage{
find_breaks_q5(min_val, max_val)
}
\arguments{
\item{min_val}{<\code{numeric}>}

\item{max_val}{<\code{numeric}>}
}
\value{
Returns a numeric vector with pretty q5 break values.
}
\description{
Find pretty q5 breaks
}
