% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{find_breaks_quintiles_q5}
\alias{find_breaks_quintiles_q5}
\title{Find pretty q5 breaks}
\usage{
find_breaks_quintiles_q5(min_val, max_val, dist)
}
\arguments{
\item{min_val}{<\code{numeric}>}

\item{max_val}{<\code{numeric}>}

\item{dist}{<\code{numeric}> Distribution (numerics) with no NAs.}
}
\value{
Returns a numeric vector with pretty q5 break values. It ensures that
each break point is different by adjusting the rounding base dynamically
based on the difference between each quantile and its predecessor. The while
loop ensures that if a new break point is the same as the last one after
rounding, the base for rounding will be decreased, increasing the precision
until a unique break point is found. Note that this might lead to break
points with a high level of precision, especially in data sets with many
similar values.
}
\description{
Find pretty q5 breaks
}
