% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_area}
\alias{get_area}
\title{Compute geometric measurements and drop units}
\usage{
get_area(geometry, ...)
}
\arguments{
\item{geometry}{<\code{sf}, \code{sfc} or \code{sfg}> A column of geometries.}

\item{...}{passed on to \code{\link[sf]{st_area}}}
}
\value{
A numeric vector of shape areas.
}
\description{
Compute geometric measurements and drop units
}
