% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avail_scale_combinations.R
\name{get_avail_scale_combinations}
\alias{get_avail_scale_combinations}
\title{Get Available Scale Combinations}
\usage{
get_avail_scale_combinations(scales_sequences, avail_scales)
}
\arguments{
\item{scales_sequences}{<\code{list}> A list of scales sequences representing the
hierarchical ordering of scales on an auto-zoom.}

\item{avail_scales}{<\verb{character vector}> A vector of character containing
all the scales at which the data of the module is accessible.}
}
\value{
A character vector of available scale combinations.
}
\description{
This function retrieves available scale combinations based on the provided
scale sequences and a character vector of scales for which the data of the page
is available. It filters out smaller scales that would normally show DA values
(like streets/buildings, so we keep scales sequences containing buildings) and
returns the combination of scales usable on a particular page.
}
