% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{get_breaks_q3}
\alias{get_breaks_q3}
\title{Get q3 break values}
\usage{
get_breaks_q3(df, vars)
}
\arguments{
\item{df}{<\code{data.frame}> Contains all columns in \code{vars} as well as their
q3s. The result of \code{\link[cc.buildr]{add_q3}}.}

\item{vars}{<\verb{vector of character}> Contains all variable names from which
to add q3s. Must fit with a name in \code{df}.}
}
\value{
A data.frame where each column in a var, and the rows are the q3
breaks.
}
\description{
Get q3 break values
}
