% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{get_breaks_q5}
\alias{get_breaks_q5}
\title{Get q5 break values}
\usage{
get_breaks_q5(df, vars)
}
\arguments{
\item{df}{<\code{data.frame}> Contains all columns in \code{vars}.}

\item{vars}{<\verb{vector of character}> Contains all variable names from q5s should
be calculated. Must fit with a name in \code{df}.}
}
\value{
A data.frame where each column in a var, and the rows are the q3
}
\description{
Get q5 break values
}
