% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_census_cc.R
\name{get_census_cc}
\alias{get_census_cc}
\title{Retrieve census data and filter it by the master polygon}
\usage{
get_census_cc(
  master_polygon,
  census_dataset,
  regions,
  level,
  var_select = c(CTUID = "CT_UID", CSDUID = "CSD_UID", name = "name", population =
    "Population", households = "Households"),
  crs,
  format = TRUE,
  cartographic = FALSE,
  area_threshold = 0.05
)
}
\arguments{
\item{master_polygon}{<\code{sfc_MULTIPOLYGON}>. Unioned multipolygon of all the
geometries for which census data must be gathered.}

\item{census_dataset}{<\code{character}> The dataset to query for available
regions, e.g. \code{"CA16"}.}

\item{regions}{<\verb{named list}> A named list of census regions to retrieve.
Names must be valid census aggregation levels. Preferably a whole province
to make sure all geometries present in the <\code{master_polygon}> is added.
e.g. \code{list(PR = 24)} for Montreal.}

\item{level}{<\code{character}> The census aggregation level to retrieve. One of \code{"CMA"},
\code{"CD"}, \code{"CT"}, \code{"DA"}, \code{"DB"}, ...}

\item{var_select}{<\verb{named character vector}> Variables to be selected from the
\code{\link[cancensus]{get_census}} cancensus function call, and renamed. Default
to \code{c("CTUID" = "CT_UID", "CSDUID" = "CSD_UID", "name" = "name",
"population" = "Population", "households" = "Households")}.}

\item{crs}{<\code{numeric}> EPSG coordinate reference system to be assigned, e.g.
\code{32618} for Montreal.}

\item{format}{<\code{logical}> Indicating if the census call should be filter
by the <\code{master_polygon}> or not.}

\item{cartographic}{<\code{logical}> Should the census geometries get switched
to the cartographic one (the one which spans over water)?}

\item{area_threshold}{<\code{numeric}> How much of a feature should be present
in the master polygon to include it in the dataset? Defaults to 0.05. In
some cases, if a DA is small on the land but covers lots of water, this
threshold must be taken down.}
}
\value{
An sf dataframe of a census scale filtered by the master polygon.
}
\description{
Retrieve census data and filter it by the master polygon
}
