% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_census_vectors_details.R
\name{get_census_vectors_details}
\alias{get_census_vectors_details}
\title{Get details of the census vectors, used for data export}
\usage{
get_census_vectors_details(census_vectors = cc.data::census_vectors)
}
\arguments{
\item{census_vectors}{<\verb{character vector}> Data variables that were added
to the scales. By default, all: \code{\link[cc.data]{census_vectors}}. Look
at the \code{\link[cc.data]{list_census_vectors}} to view all
variables explained.}
}
\value{
The wanted subset of \code{\link[cc.data]{census_vectors_details}}
}
\description{
Get details of the census vectors, used for data export
}
