% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cmhc_zones.R
\name{get_cmhc_zones}
\alias{get_cmhc_zones}
\title{Get CMHC zones in a certain CMA}
\usage{
get_cmhc_zones(cancensus_region, crs = NULL)
}
\arguments{
\item{cancensus_region}{A list containing a named vector. It must be a valid
\code{regions} call argument to the \code{\link[cancensus]{get_census}} function,
e.g. \code{list(CMA = 24462)} for Montreal.}

\item{crs}{<\code{numeric}> Optional. EPSG coordinate reference system to be
assigned, e.g. \code{32618} for Montreal. Optional, defaults to the UTM zone
retrieved by the centroid of the master_polygon created. If not supplied, it
is derived from the centroid of all the geos through a simple mathematical
approach.}
}
\value{
An sf data.frame of two columns: name of the zone and geometry.
}
\description{
Get CMHC zones in a certain CMA
}
