% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heroku.R
\name{heroku_restart_dyno}
\alias{heroku_restart_dyno}
\title{Restart dyno for the provided app}
\usage{
heroku_restart_dyno(app_name, dyno)
}
\arguments{
\item{app_name}{<\code{character}> Name of the application.}

\item{dyno}{<\code{character}> Name of the dyno.}
}
\value{
Opens a terminal asking to login, then restarts the dyno.
}
\description{
Restart dyno for the provided app
}
