% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate_fast_additive_sum}
\alias{interpolate_fast_additive_sum}
\title{Fast Summation for interpolation}
\usage{
interpolate_fast_additive_sum(col_name, data, id_col, weight_col)
}
\arguments{
\item{col_name}{<\code{character}> A string representing the name of the column
to be summed in the input data frame.}

\item{data}{<\code{data.frame}> A data frame containing the data to be summed. It
must have columns corresponding to the specified id_col and col_name.}

\item{id_col}{<\code{character}> A string representing the name of the ID column
in the input data frame.}

\item{weight_col}{<\code{character}> A string representing the name of the weight
column in the input data frame.}
}
\value{
A data frame with two columns: the ID column (with the same name as
in the input data frame) and a column containing the summed values, named
after the input col_name.
}
\description{
This function computes the the summation for each group within a
data frame. The data frame must have columns for ID and the column to be
summed. The output is a data frame containing the IDs and their corresponding
summed values.
}
