% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate_from_area}
\alias{interpolate_from_area}
\title{Interpolate ADDITIVE variables using area}
\usage{
interpolate_from_area(
  to,
  DA_table,
  additive_vars = c("population", "households"),
  crs
)
}
\arguments{
\item{to}{<\verb{sf data.frame}> Table for which additive data must be
interpolated}

\item{DA_table}{<\verb{sf data.frame}> A \code{DA} sf data.frame from which
population and households will be interpolated.}

\item{additive_vars}{<\verb{vector of character}> Columns from \code{DA_table} from which
data should be interpolated to the \code{to} data.frame. Must be ADDITIVE variables,
e.g. population and households (default).}

\item{crs}{<\code{numeric}> EPSG coordinate reference system to be assigned, e.g.
\code{32618} for Montreal.}
}
\value{
Returns the \code{to} data.frame with the added or modified columns that
have been interpolated from the \code{DA_table}.
}
\description{
Interpolate ADDITIVE variables using area
}
