% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_over_scales}
\alias{map_over_scales}
\title{Map over the depth of the list of scales}
\usage{
map_over_scales(all_scales, fun)
}
\arguments{
\item{all_scales}{<\verb{named list}> A named list of scales. The first level is
the geo, and the second is the scales.}

\item{fun}{<\code{function}> A function to run on all scales. Possible arguments
are \code{geo} (names of regions), \code{scales} (scales data.frame), 3
\code{scale_name} (names of scales inside a region), \code{scale_df} (data.frame, single
scale)}
}
\value{
Returns the all_scales list that is fed, with the \code{fun} ran
on all the scales.
}
\description{
As we follow a deep list of scales in regions, we can use this \code{map_over_scales()}
helper function to iterate functions on all the scales.
}
