% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_zoom_levels.R
\name{map_zoom_levels_create_all}
\alias{map_zoom_levels_create_all}
\title{Create all basic map zoom levels}
\usage{
map_zoom_levels_create_all(
  all_tables,
  zoom_levels = list(first = 0, CT = 10.5, DA = 12.5, building = 15.5)
)
}
\arguments{
\item{all_tables}{<\verb{named list}> A named list of character. The names are
the \code{region} e.g. CMA, island, city, ... and the vectors are all the
scales available in these geo e.g. CSD, CT, DA, building, ...}

\item{zoom_levels}{<\verb{named list}> A numeric list with names. Every scale
that is not located first in the region's list must be part of this list,
with a numeric applied to it (at which zoom level will the user switch between
one scale to another on auto-zoom). The first element of every region's list
of scales has the name \code{first} in that list, which will always be defaulted
to 0.}
}
\value{
Returns a named list of all basic map zoom levels.
}
\description{
These zoom levels are used used to inform the zoom slider. The default
goes: first scale in the region, then \code{CT}, \code{DA} and \code{building} or \code{street}.
}
