% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_zoom_levels.R
\name{map_zoom_levels_create_custom}
\alias{map_zoom_levels_create_custom}
\title{Create a custom zoom level}
\usage{
map_zoom_levels_create_custom(
  map_zoom_levels,
  all_tables,
  region,
  suffix,
  content
)
}
\arguments{
\item{map_zoom_levels}{<\verb{named list}> The previously created zoom levels
using \code{\link[cc.buildr]{map_zoom_levels_create_all}}}

\item{all_tables}{<\verb{named list}> A named list of character. The names are
the \code{region} e.g. CMA, island, city, ... and the vectors are all the
scales available in these geo e.g. CSD, CT, DA, building, ...}

\item{region}{<\code{character}> The region for which this new custom map zoom
level should be for, e.g. \code{CMA}.}

\item{suffix}{<\code{character}> The name that should be appended to the name
of the map zoom level, e.g. \code{max_CT}. The full name would then be:
\code{map_zoom_levels_max_CT}.}

\item{content}{<\code{character}> The content of the zoom level, e.g.
\code{c("CSD" = 0, "CT" = 10.5)}.}
}
\value{
The same list as \code{map_zoom_levels} with the new custom zoom level
appended.
}
\description{
Create a custom zoom level
}
