% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_zoom_levels.R
\name{map_zoom_levels_save}
\alias{map_zoom_levels_save}
\title{Save all map zoom levels in a \code{.qsm} file}
\usage{
map_zoom_levels_save(data_folder = "data/", map_zoom_levels)
}
\arguments{
\item{data_folder}{<\code{character}> Where the \code{.sqlite} databases should be
written to. Defaults to \verb{data/}.}

\item{map_zoom_levels}{<\verb{named list}> The previously created zoom levels
using \code{\link[cc.buildr]{map_zoom_levels_create_all}} and
\code{\link[cc.buildr]{map_zoom_levels_create_custom}}}
}
\value{
Returns  nothing if ran successfully.
}
\description{
Save all map zoom levels in a \code{.qsm} file
}
