% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge}
\alias{merge}
\title{Merge two tibbles and keep the output a tibble}
\usage{
merge(x, y, by = intersect(names(x), names(y)), by.x = by, by.y = by, ...)
}
\arguments{
\item{x}{<\code{data.frame}> data frame, or object to be coerced to one.}

\item{y}{<\code{data.frame}> data frame, or object to be coerced to one.}

\item{...}{arguments to be passed to \code{\link[base]{merge}}.}
}
\value{
Merged tibbles as a tibble and keep sf class if was present
}
\description{
Merge two tibbles and keep the output a tibble
}
