% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paqes_regions.R
\name{pages_regions}
\alias{pages_regions}
\title{Pages Regions Function}
\usage{
pages_regions(modules, regions_dictionary)
}
\arguments{
\item{modules}{<\code{list}> Dataframe of modules.}

\item{regions_dictionary}{<\code{data.frame}> The regions dictionary,}
}
\value{
<\code{list}> An SVM object with the \code{modules} element updated to include
\code{regions}, which are the filtered regions based on the data coverage criteria.
}
\description{
This function identifies and filters regions based on the availability and
coverage of data in different scales for each module in a given SVM object.
It evaluates each region's data presence and only keeps regions that have a
high enough percentage of features filled with data.
}
