% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_explorer.R
\name{placeex_main_card}
\alias{placeex_main_card}
\title{Produce the place explorer main card}
\usage{
placeex_main_card(scales, DA_table, region_DA_IDs, crs, regions_dictionary)
}
\arguments{
\item{scales}{<\code{list}> Lists of spatial features dataframes with regions and
scales filled with at minimum census and canale data. Usually is
\code{scales_variables_modules$scales}.}

\item{DA_table}{<\verb{sf data.frame}> Tables containing all the dissemination areas
of the zone under study. Usually is \code{census_scales$DA}}

\item{region_DA_IDs}{<\verb{character vector}> Vector of all the dissemination area's
identifiers in the zone under study. Usually is \code{census_scales$DA$ID}}

\item{crs}{<\code{numeric}> EPSG coordinate reference system to be assigned, e.g.
\code{32617} for Toronto.}

\item{regions_dictionary}{<\code{data.frame}> The regions dictionary built using
\code{\link[cc.buildr]{regions_dictionary}}. Will be used to filter out scales
for which data should not be calculated.}
}
\value{
Lists containing all the main card indicators. Sub lists of regions
and scales for data values.
}
\description{
Produce the place explorer main card
}
