% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_explorer.R
\name{placeex_main_card_data}
\alias{placeex_main_card_data}
\title{Produce the place explorer main card}
\usage{
placeex_main_card_data(
  scales_dictionary,
  DA_table,
  region_DA_IDs,
  crs,
  regions_dictionary,
  ignore_scales = c("building", "street", "grd30", "grd60", "grd120", "grd300", "grd600"),
  first_scales
)
}
\arguments{
\item{scales_dictionary}{<\code{list}> Scales dictionary}

\item{DA_table}{<\verb{sf data.frame}> Tables containing all the dissemination areas
of the zone under study. Usually is \code{census_scales$DA}}

\item{region_DA_IDs}{<\verb{character vector}> Vector of all the dissemination area's
identifiers in the zone under study. Usually is \code{census_scales$DA$ID}}

\item{crs}{<\code{numeric}> EPSG coordinate reference system to be assigned, e.g.
\code{32617} for Toronto.}

\item{regions_dictionary}{<\code{data.frame}> The regions dictionary built using
\code{\link[cc.buildr]{regions_dictionary}}. Will be used to filter out scales
for which data should not be calculated.}

\item{first_scales}{<\verb{character vector}> Scales for which place explorer should
be calculated}
}
\value{
Lists containing all the main card indicators. Sub lists of regions
and scales for data values.
}
\description{
Produce the place explorer main card
}
