% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_explorer.R
\name{placeex_main_card_prep_output_en}
\alias{placeex_main_card_prep_output_en}
\title{Prepare data for a title card (english)}
\usage{
placeex_main_card_prep_output_en(
  data,
  dict,
  region,
  scale,
  select_id,
  regions_dictionary,
  scales_dictionary
)
}
\arguments{
\item{data}{<\code{list}> Contains data about the indicator in its region
and scale, e.g. \code{pe_main_card$main_card_data$no2$CMA$CSD}.}

\item{dict}{<\code{data.frame}> Row from the dictionary for the data
(title, text, etc.), e.g.
\code{pe_main_card$main_card_dict[pe_main_card$main_card_dict$name == "no2", ]}.}

\item{region}{<\code{character}> The geographic region.}

\item{scale}{<\code{data.frame}> The scale, e.g. \code{DA}, \code{CT}, ...}

\item{select_id}{<\code{character}> The ID of the selected area.}

\item{regions_dictionary}{<\code{data.frame}> A dictionary containing information
about the regions.}

\item{scales_dictionary}{<\code{data.frame}> The scales dictionary built using
\code{\link[cc.buildr]{build_census_scales}}}
}
\value{
A list of information about the data, including the data value,
data date, data rank (in text), and color category (1-5).
}
\description{
This function prepares data for a title card for a specific region and indicator.
It returns information about the data value, data date, data rank and color.
}
