% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_explorer.R
\name{placeex_main_card_rmd}
\alias{placeex_main_card_rmd}
\title{Pre-process all the possible Rmds}
\usage{
placeex_main_card_rmd(
  scales,
  pe_main_card_data,
  regions_dictionary,
  scales_dictionary,
  out_folder = "www/place_explorer/",
  lang = "en",
  tileset_prefix,
  mapbox_username = "sus-mcgill"
)
}
\arguments{
\item{scales}{<\code{list}> Lists of spatial features dataframes with regions and
scales filled with at minimum census and canale data. Usually is
\code{scales_variables_modules$scales}.}

\item{pe_main_card_data}{<\code{list}> Data and dictionary necessary to knit the
rmds. The output of \code{\link[cc.buildr]{placeex_main_card_data}}.}

\item{regions_dictionary}{<\code{data.frame}> The regions dictionary built using
\code{\link[cc.buildr]{regions_dictionary}}. Will be used to filter out scales
for which data should not be calculated.}

\item{scales_dictionary}{<\code{data.frame}> The scales dictionary built using
\code{\link[cc.buildr]{build_census_scales}}}

\item{out_folder}{<\code{character}> Folder where all the place explorer HTML
documents should be saved. Must be in the www folder to be used by the
shinyapp. Defaults to \verb{www/place_explorer/}}

\item{lang}{<\code{character}> If the Curbcut instance is bilingual, we can use
\code{c("en", "fr")} and HTML will be built for both languages. Defaults to \code{"en"}.}

\item{tileset_prefix}{<\code{character}> Prefix attached to every tile source and
created and published tileset. Should correspond to the Curbcut city, e.g. \code{mtl}.}

\item{mapbox_username}{<\code{character}> Mapbox account username.}
}
\value{
Returns nothing if successful. All place explorer possibilities are
saved in the \code{out_folder}.
}
\description{
Pre-process all the possible Rmds
}
