% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{post_processing}
\alias{post_processing}
\title{Post process scales}
\usage{
post_processing(scales_variables_modules)
}
\arguments{
\item{scales_variables_modules}{<\verb{named list}> A list of length three.
The first is all the scales, the second is the variables table, and the
third is the modules table.}
}
\value{
A list containing the post-processed scales, variables and modules
}
\description{
This function casts the geometry of scales to the right type, and reorders
the columns so that IDs and names are first.
}
