% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{provinces_pbf}
\alias{provinces_pbf}
\title{Province's geometries and pbf download link file}
\format{
\subsection{\code{provinces_pbf}}{

An sf data frame with 10 rows and 3 columns:
\describe{
\item{name}{The province's name}
\item{link}{Link from where to download the pbf file}
\item{geometry}{MULTIPOLYGON of the province}
}
}
}
\usage{
provinces_pbf
}
\description{
Province's geometries and pbf download link file
}
\keyword{datasets}
