% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ready_to_use.R
\name{ready_to_use_canbics}
\alias{ready_to_use_canbics}
\title{Add a ready to use Can-BICS data and module}
\usage{
ready_to_use_canbics(scales_variables_modules, crs)
}
\arguments{
\item{scales_variables_modules}{<\verb{names list}> A list of length three.
The first is all the scales, the second is the variables table, and the
third is the modules table.}

\item{crs}{<\code{numeric}> EPSG coordinate reference system to be assigned, e.g.
\code{32617} for Toronto.}
}
\value{
A list of length 3, similar to the one fed to
\code{scales_variables_modules} with the Can-BICS variable added, its addition
in the variables table and the module table.
}
\description{
Add a ready to use Can-BICS data and module
}
