% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reconstruct_all_tables}
\alias{reconstruct_all_tables}
\title{Reconstruct all_tables}
\usage{
reconstruct_all_tables(all_scales)
}
\arguments{
\item{all_scales}{<\verb{named list}> A named list of scales. The first level is
the geo, and the second is the scales.}
}
\value{
A list of vectors of characters. The first level of the list is the
geo, and the second is all the scales available in that particular geo.
}
\description{
As all the data arrange in a named list of geos > scales was created using
a list of character vector previously named \code{all_tables}, this function
allows to recreate the \code{all_tables} list of vectors using the data of
\code{all_scales}.
}
