% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions_dictionary.R
\name{regions_dictionary}
\alias{regions_dictionary}
\title{Build the regions dictionary}
\usage{
regions_dictionary(all_tables, region, name, to_compare, pickable)
}
\arguments{
\item{all_tables}{<\verb{named list}> Named list of regions and their scales within,
ordered in priority.}

\item{region}{<\verb{vector of character}> A vector of characters used to identify
which large geometry the user is interested in, e.g.
\code{c("CMA", "city", "island")}.}

\item{name}{<\verb{named character vector}> Named with their corresponding geo value.
Same length as the vector fed in the \code{geo} argument. Used to name the geo
when the user will want to change which large geometry they are interested in. e.g.
\code{c("Metropolitan Area", "City of Montreal", "Island of Montreal")}.}

\item{to_compare}{<\verb{named character vector}> Named with their corresponding geo value.
Same length as the vector fed in the \code{geo} argument. The same length as
the vector fed in the \code{geo} argument. Used at the end of dynamically generated
text, when a particular zone is compared with the whole geography. 'x DA
has a higher value than y\% of DAs in the Montreal region'.
e.g. \code{c("in the Montreal region", "in the City of Montreal",
"on the island of Montreal")}}

\item{pickable}{<\verb{named vector of logical}> Will the user be able to select this
scale as a default all over the platform?}
}
\value{
Returns the same vectors fed arranged in a data.frame ordered in
priorty.
}
\description{
Build the regions dictionary
}
