% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{regions_dictionary_add_scales}
\alias{regions_dictionary_add_scales}
\title{Add \code{scales} column to the regions_dictionary}
\usage{
regions_dictionary_add_scales(regions_dictionary, region_dict_scales)
}
\arguments{
\item{regions_dictionary}{<\code{data.frame}> The regions dictionary built using
\code{\link[cc.buildr]{regions_dictionary}}. Will be used to filter out scales
for which data should not be calculated.}

\item{region_dict_scales}{<\verb{named list}> Named list of the IDs of all scales
that fit within every region. One of the output of \code{\link[cc.buildr]{scales_consolidated}}}
}
\value{
The same \code{regions_dictionary} fed with the \code{scales} column added.
}
\description{
Add \code{scales} column to the regions_dictionary
}
