% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reorder_columns}
\alias{reorder_columns}
\title{Reorder columns as they are expected in the workflow}
\usage{
reorder_columns(scale_df)
}
\arguments{
\item{scale_df}{<\code{data.frame}> A \code{scale} data.frame for which columns
should be reordered.}
}
\value{
Returns the same list that is fed, with columns re-ordered.
}
\description{
Reorder columns as they are expected in the workflow
}
