% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stories.R
\name{resize_image}
\alias{resize_image}
\title{Resize images in a directory (Mostly used for stories)}
\usage{
resize_image(folder = "www/stories/photos", max_size_in_MB = 1)
}
\arguments{
\item{folder}{<\code{character}> A string indicating the path to the directory with
images. The default is "www/stories/photos".}

\item{max_size_in_MB}{<\code{numeric}> Value defining the maximum allowed size for
the images in Megabytes (MB). Images larger than this size will be
rescaled. The default is 1 MB.}
}
\value{
NULL. The function works by side effects, resizing the images and
overwriting the original files in their original directory.
}
\description{
This function reduces the size of images in a directory if they exceed a
specified maximum size. The function works recursively and supports the
.jpg, .jpeg, and .png file formats. The reduction is done by rescaling the
images to 50\% of their original dimensions iteratively until they fall
below the maximum specified size.
}
