% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_geocode.R
\name{rev_geocode_BC}
\alias{rev_geocode_BC}
\title{Reverse geocode for points in British Columbia}
\usage{
rev_geocode_BC(point_sf)
}
\arguments{
\item{point_sf}{< \code{sfc_POINT}> Point sf, e.g. a centroid of a building}
}
\value{
The address retrieved from the BC Geocoder service
}
\description{
Reverse geocode for points in British Columbia
}
