% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_geocode.R
\name{rev_geocode_OSM}
\alias{rev_geocode_OSM}
\title{Reverse geocode using OSM}
\usage{
rev_geocode_OSM(point_sf)
}
\arguments{
\item{point_sf}{< \code{sfc_POINT}> Point sf, e.g. a centroid of a building}
}
\value{
The address retrieved from the OSM service
}
\description{
Reverse geocode using OSM
}
