% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_geocode.R
\name{rev_geocode_QC}
\alias{rev_geocode_QC}
\title{Reverse geocode for a point in Quebec}
\usage{
rev_geocode_QC(point_sf)
}
\arguments{
\item{point_sf}{< \code{sfc_POINT}> Point sf, e.g. a centroid of a building}
}
\value{
The address retrieved from the Adresses Québec service
}
\description{
Reverse geocode for a point in Quebec
}
