% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_geocode_sf.R
\name{rev_geocode_sf}
\alias{rev_geocode_sf}
\title{Reverse geocode spatial features}
\usage{
rev_geocode_sf(master_polygon, sf_df, province_code, crs)
}
\arguments{
\item{master_polygon}{<\code{sfc_MULTIPOLYGON}> Unioned multipolygon of all the
geometries for which sf's addresses should be gathered. The National
Open Database of Addresses will then be spatially filter using it, reducing
computation time of \code{\link[sf]{st_nearest_feature}}}

\item{sf_df}{<\verb{sf data.frame}> Any data.frame with spatial features}

\item{province_code}{<\code{character}> Province code used to download the
Open Database of Addresses. One of \code{"AB"}, \code{"BC"}, \code{"MB"},
\code{"NB"}, \code{"NT"}, \code{"NS"}, \code{"ON"}, \code{"PE"}, \code{"QC"} or
\code{"SK"}.}

\item{crs}{<\code{numeric}> EPSG coordinate reference system to be assigned, e.g.
\code{32618} for Montreal.}
}
\value{
The same sf data.frame fed, with the \code{name} column populated
with the reverse geocoded addresses.
}
\description{
This geocoding works in three steps. The province database of addresses is
first downloaded from the National Open Database of Addresses, with URLs to
download the zip files available in \code{\link[cc.buildr]{addresses_db_links}}.
For the missing addresses, we use the province's reverse geolocating service
(at the moment only available for QC and BC). Third, we reverse geocode using
the OSM service. The reverse geocoding is done using the centroid of every
spatial feature.
}
