% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rough_rank}
\alias{rough_rank}
\title{Reimplemntation of dplyr::ntile in base R}
\usage{
rough_rank(x, n)
}
\arguments{
\item{x}{<\verb{numeric vector}> A vector of values to rank. Missing values are
left as is. If you want to treat them as the smallest or largest values,
replace with Inf or -Inf before ranking.}

\item{n}{<\code{numeric}> Number of groups to split up into.}
}
\value{
A rough rank, which breaks the input vector into n buckets. The size
of the buckets may differ by up to one, larger buckets have lower rank.
}
\description{
Reimplemntation of dplyr::ntile in base R
}
