% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ready_to_use.R
\name{ru_ndvi}
\alias{ru_ndvi}
\title{Add a ready to use Greenness (NDVI) data and module}
\usage{
ru_ndvi(scales_variables_modules, region_DA_IDs, crs)
}
\arguments{
\item{scales_variables_modules}{<\verb{named list}> A list of length three.
The first is all the scales, the second is the variables table, and the
third is the modules table.}

\item{region_DA_IDs}{<\verb{character vector}> All the current census'
DA IDs present in the region. Only those will be extracted from the database
to do interpolation.}

\item{crs}{<\code{numeric}> EPSG coordinate reference system to be assigned, e.g.
\code{32617} for Toronto.}
}
\value{
A list of length 3, similar to the one fed to
\code{scales_variables_modules} with the Can-BICS variable added, its addition
in the variables table and the module table.
}
\description{
Add a ready to use Greenness (NDVI) data and module
}
