% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ready_to_use.R
\name{ru_vac_rate}
\alias{ru_vac_rate}
\title{Add a ready to use Vacancy Rate data and module}
\usage{
ru_vac_rate(scales_variables_modules, crs, geo_uid)
}
\arguments{
\item{scales_variables_modules}{<\verb{names list}> A list of length three.
The first is all the scales, the second is the variables table, and the
third is the modules table.}

\item{crs}{<\code{numeric}> EPSG coordinate reference system to be assigned, e.g.
\code{32617} for Toronto.}

\item{geo_uid}{<\code{numeric}> Cancensus CMA code, which can be found using
\code{\link[cancensus]{list_census_regions}}.}
}
\value{
A list of length 3, similar to the one fed to
\code{scales_variables_modules} with the CMHC's vacancy rate variables added,
their addition in the variables table and the module table.
}
\description{
Add a ready to use Vacancy Rate data and module
}
