% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data_files.R
\name{save_all_scales_qs}
\alias{save_all_scales_qs}
\title{Save all scale tables in QS format}
\usage{
save_all_scales_qs(scales_dictionary, data_folder = "data/")
}
\arguments{
\item{scales_dictionary}{<\verb{named list}> Dictionary of scales}

\item{data_folder}{<\code{character}> Where the \code{.qs} files should be
written to. Defaults to \verb{data/}.}
}
\value{
An invisible NULL value.
}
\description{
This function saves all scale tables in QS format in the specified data folder.
It first drops the geometry of other scales and creates a list of tables for all
scales, then creates the folders for each region and scale, and finally saves the
tables in the database.
}
