% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data_files.R
\name{save_all_scales_sqlite}
\alias{save_all_scales_sqlite}
\title{Save every scales in their own SQLite database}
\usage{
save_all_scales_sqlite(
  data_folder = "data/",
  all_scales,
  variables,
  scales_to_drop = c("building", "street")
)
}
\arguments{
\item{data_folder}{<\code{character}> Where the \code{.sqlite} databases should be
written to. Defaults to \verb{data/}.}

\item{all_scales}{<\verb{named list}> A named list of sf data.frame
containing all scales listed with their regions, normally
\code{scales_variables_modules$scales}.}

\item{variables}{<\code{data.frame}> The \code{variables} data.frame, normally
\code{scales_variables_modules$variables}.}

\item{scales_to_drop}{<\verb{character vector}> Scales that shouldn't have their
own SQLite database. Defaults to \code{c("building", "street")}.}
}
\value{
Returns an error or nothing if ran successfully. Every existing region-geo
combination is a new SQLite db, and every variable is a table saved in each
of the db.
}
\description{
Save every scales in their own SQLite database
}
