% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data_files.R
\name{save_bslike_sqlite}
\alias{save_bslike_sqlite}
\title{Save all buildings-street-like dataset in a SQLite}
\usage{
save_bslike_sqlite(
  scale_chr,
  path = sprintf("data/\%s.sqlite", scale_chr),
  all_scales
)
}
\arguments{
\item{scale_chr}{<\code{character}> The name of the scale, e.g. "building".}

\item{path}{<\code{character}> Path where to save the \code{.sqlite}. Defaults to
\verb{data/'scale_chr'.sqlite}}

\item{all_scales}{<\verb{named list}> A named list of sf data.frame
containing all scales listed with their regions, normally
\code{scales_variables_modules$scales}.}
}
\value{
Returns an error or nothing if ran successfully. All existing \code{scale_chr}
data.frame in the fed \code{all_scales} are saved in the created \code{.sqlite}.
}
\description{
The function is used for any scales that ressembles to buildings or street, in
that they are very large in the amount of observations and most of the time,
the data attached to it on Curbcut will be the one of the dissemination areas.
}
