% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data_files.R
\name{save_buildings_sqlite}
\alias{save_buildings_sqlite}
\title{Save all buildings dataset in a SQLite}
\usage{
save_buildings_sqlite(path = "data/building.sqlite", all_scales)
}
\arguments{
\item{path}{<\code{character}> Path where to save the \code{.sqlite}. Defaults to
\code{data/building.sqlite}}

\item{all_scales}{<\verb{named list}> A named list of sf data.frame
containing all scales listed with their regions, normally
\code{scales_variables_modules$scales}.}
}
\value{
Returns an error or nothing if ran successfully. All existing \code{building}
data.frame in the fed \code{all_scales} are saved in the created \code{.sqlite}.
}
\description{
Save all buildings dataset in a SQLite
}
