% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_postal_codes.R
\name{save_postal_codes}
\alias{save_postal_codes}
\title{Save the Postal Codes Data}
\usage{
save_postal_codes(region_DA_IDs, overwrite = FALSE, data_folder = "data/")
}
\arguments{
\item{region_DA_IDs}{<\verb{character vector}> All the current census'
DA IDs present in the region. Only those will be extracted from the database
to do interpolation.}

\item{overwrite}{<\code{logical}> If TRUE, will overwrite the existing file if it exists.
Default is FALSE.}

\item{data_folder}{<\code{character}> The folder path where the postal codes file will be saved.
Default is "data/".}
}
\value{
Invisible NULL. The function is used for its side effect of saving a
file and does not return a value.
}
\description{
Saves the postal codes data for a specified region into a quickstore (qs)
file. If the file already exists, it can optionally overwrite it.
}
