% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndvi_grids.R
\name{scales_dictionary_ndvi}
\alias{scales_dictionary_ndvi}
\title{Append NDVI scales to the scales dictionary.}
\usage{
scales_dictionary_ndvi(scales_dictionary)
}
\arguments{
\item{scales_dictionary}{<\code{data.frame}> A dictionary to which NDVI scales will be added.}
}
\value{
Updated scales dictionary with NDVI scale entries.
}
\description{
This function adds multiple NDVI scale entries to a given dictionary. Each
entry contains details about a specific scale, such as its singular and
plural forms, slider title, and descriptive text.
}
