% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{scales_greater_than}
\alias{scales_greater_than}
\title{Determine which scales are greater in size than the base scale}
\usage{
scales_greater_than(base_scale, all_scales, crs)
}
\arguments{
\item{base_scale}{<\code{data.frame}> A spatial data frame representing the base
scale to be compared.}

\item{all_scales}{<\verb{named list}> A list of spatial data frames representing
the scales to be compared.}

\item{crs}{<\code{numeric}> EPSG coordinate reference system to be assigned, e.g.
\code{32618} for Montreal.}
}
\value{
A character vector containing the names of the scales that are greater
in size than the base scale.
}
\description{
This function compares the average size of a base scale to other scales,
transforming them to the same coordinate reference system (CRS) for an accurate
comparison. It then returns the names of the scales which are greater than
(or very close to) the base scale, including itself..
}
