% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spatial_filtering}
\alias{spatial_filtering}
\title{Spatial filtering function to keep polygons with at least x of their area
in the greater boundary.}
\usage{
spatial_filtering(
  df,
  crs,
  master_polygon,
  ID_col = "ID",
  area_threshold = 0.05
)
}
\arguments{
\item{df}{<\verb{sf data.frame}> The census data object to be transformed.}

\item{crs}{<\code{numeric}> EPSG coordinate reference system to be assigned, e.g.
\code{32618} for Montreal.}

\item{master_polygon}{<\code{sfc_MULTIPOLYGON}> Unioned multipolygon of the boundaries
in which to keep the df features.}

\item{ID_col}{<\code{character}> The name of the ID column to be used for filtering.
Default is "ID".}

\item{area_threshold}{<\code{numeric}> The minimum percentage of area in the master polygon.
Default is 0.05, as a low percentage can be use for the fact that water can sometimes
be a large part of the polygons.}
}
\value{
A subset of the original data frame that meets the filtering condition.
}
\description{
Spatial filtering function to keep polygons with at least x of their area
in the greater boundary.
}
