% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_scale.R
\name{split_scale}
\alias{split_scale}
\title{Split a (probably census) layer with a set of boundaries to cut the first
layer with}
\usage{
split_scale(
  destination,
  cutting_layer,
  DA_table,
  destination_pct_threshold = 0.95,
  buffer_around_cutting_layer = 100,
  sampled_points_voronoi = buffer_around_cutting_layer * 500,
  crs
)
}
\arguments{
\item{destination}{<\verb{sf data.frame}> Probably a census layer, which is
the destination layer, for which some polygons should be cut using the
polygons of the \code{cutting_layer}.}

\item{cutting_layer}{<\verb{sf data.frame}> An sf data.frame of subdivisions of one
or more polygons of the destination data.frame, e.g. for the City of Montreal,
the subdivisions is a dataframe of boroughs. The sf data.frame must have 3
columns: \code{name}, \code{type} and \code{geometry}.
The \code{name} is the name of the borough/neighbourhood/zone e.g.
"Le Plateau Mont-Royal", and \code{type} is what should be used as \code{"name_2"},
e.g. "Borough" (for the display name: Borough of Le Plateau Mont-Royal).}

\item{DA_table}{<\verb{sf data.frame}> A \code{DA} sf data.frame from which
population and households will be interpolated.}

\item{destination_pct_threshold}{<\code{numeric}> What is the threshold for which
an existing feature in the destination data.frame should be removed. Default to
0.95. If 95\% of the area of a destination feature fits in the \code{cutting_layer},
then it should be swapped for the polygons in the \code{cutting_layer}.}

\item{buffer_around_cutting_layer}{<\code{numeric}> Needed if the boundaries of
the \code{cutting_layer} are too short of the removed destination features. The
\code{cutting_layer} should be expanded: defaults to 100m. This will NOT make
the \code{cutting_layer} overlap on non-removed features of the destination dataframe.}

\item{sampled_points_voronoi}{<\code{numeric}> Used in the same cases of
\code{buffer_around_cutting_layer}. We found that it works better if the ration
buffer:point is very high (e.g. 1:500).}

\item{crs}{<\code{numeric}> EPSG coordinate reference system to be
assigned, e.g. \code{32618} for Montreal.}
}
\value{
An sf dataframes with features subdivided.
}
\description{
Split a (probably census) layer with a set of boundaries to cut the first
layer with
}
