% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stories.R
\name{stories_add_story}
\alias{stories_add_story}
\title{Add a story to the stories table}
\usage{
stories_add_story(stories, name_id, title, preview, lon, lat)
}
\arguments{
\item{stories}{<\code{data.frame}> The \code{stories} data.frame to which add a
new row.}

\item{name_id}{<\code{character}> Name identifier for a story, lowercased and
no spaces, e.g. \code{metro_evolution}.}

\item{title}{<\code{character}> Full title of the story, e.g.
\verb{The Evolution of the Montreal Metro}}

\item{preview}{<\code{character}> Teaser for the story to appear elsewhere
in Curbcut. Usually a single sentence.}

\item{lon}{<\code{numeric}> Longitude where the story should be placed on the map.
Must be in EPSG Projection 4326 - WGS 84}

\item{lat}{<\code{numeric}> Lattitude where the story should be placed on the map.
Must be in EPSG Projection 4326 - WGS 84}
}
\value{
The same \code{stories} data.frame fed, with the added row.
}
\description{
Add a story to the stories table
}
