% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stories.R
\name{stories_atlas_mapping}
\alias{stories_atlas_mapping}
\title{Create an image atlas and mapping for a story}
\usage{
stories_atlas_mapping(
  stories,
  input_img_location = "dev/data/stories/main_img/",
  output_bandeau_location = "www/stories/bandeau_img/",
  output_atlas_location = "www/stories/image_atlas.png"
)
}
\arguments{
\item{stories}{<\code{dataframe}>A data frame containing the stories.}

\item{input_img_location}{<\code{character}> The location of the input images.
Defaults to \code{"dev/data/stories/main_img/"}}

\item{output_bandeau_location}{<\code{character}> The location to save the banner
images. Defaults to \code{"www/stories/bandeau_img.png"}}

\item{output_atlas_location}{<\code{character}> The location to save the image
atlas. Defaults to \code{"www/stories/image_atlas.png"}}
}
\value{
A list containing the mapping for the stories. This will be an input
for the stories rdeck.
}
\description{
Create an image atlas and mapping for a story
}
