% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stories.R
\name{stories_map_images}
\alias{stories_map_images}
\title{Create an image for every story, used on the map}
\usage{
stories_map_images(stories, input_img_location = "dev/data/stories/main_img/")
}
\arguments{
\item{stories}{<\code{dataframe}>A data frame containing the stories.}

\item{input_img_location}{<\code{character}> The location of the input images.
Defaults to \code{"dev/data/stories/main_img/"}}
}
\value{
A list containing the mapping for the stories. This will be an input
for the stories rdeck.
}
\description{
Create an image for every story, used on the map
}
