% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_explorer.R
\name{tilejson}
\alias{tilejson}
\title{Generate a Mapbox Tile JSON}
\usage{
tilejson(mapbox_username, tileset_prefix, tile, return_error = FALSE)
}
\arguments{
\item{mapbox_username}{<\code{character}> string representing the Mapbox username.}

\item{tileset_prefix}{<\code{character}> Prefix attached to every tileset. Should
correspond to the Curbcut city, e.g. \code{mtl}.}

\item{tile}{<\code{character}> The tile name to be fetched.}

\item{return_error}{<\code{logical}> Print the error if the tileset isn't found.}
}
\value{
A JSON list if succesfull. If missing tile, returns NULL preventing
the app from crashing. If the tile is missing and it's a _building tile,
grab the first region of the regions_dictionary and show buildings for those.
}
\description{
Given a Mapbox username, tileset prefix, and tile name, this function
generates a Mapbox Tile JSON using \code{\link[rdeck]{tile_json}}. If the
specified tile is not found, a warning message is displayed and NULL is
returned. This prevents the app from crashing.
}
