% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilesets.R
\name{tileset_buildings_texture}
\alias{tileset_buildings_texture}
\title{Upload buildings as textures for Curbcut}
\usage{
tileset_buildings_texture(building, prefix, username, access_token)
}
\arguments{
\item{building}{<\verb{sf data.frame}> A spatial object representing buildings. This
data will be transformed and used as source for a new tileset.}

\item{prefix}{<\code{character}> Prefix attached to every tile source and
created and published tileset. Should correspond to the Curbcut city, e.g. \code{mtl}.}

\item{username}{<\code{character}> Mapbox account username.}

\item{access_token}{<\code{character}> Private access token to the Mapbox account.}
}
\value{
This function does not return a value but performs side effects by
manipulating data on a Mapbox account. Specifically, it deletes a tileset and its source,
uploads a new source, creates a new tileset and its recipe, and finally, publishes
the new tileset.
}
\description{
This function first deletes a tileset and its source based on given parameters.
It then processes a building dataset to a single feature, transforms it and uploads
it as a source for a new tileset. The function also creates a recipe for the new
tileset, and subsequently creates and publishes the tileset.
}
