% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilesets.R
\name{tileset_create_recipe}
\alias{tileset_create_recipe}
\title{Create a recipe (character)}
\usage{
tileset_create_recipe(
  layer_names,
  source,
  minzoom,
  maxzoom,
  layer_size = 2500,
  simp_zoom = NULL,
  simp_value = NULL,
  fallback_simp_zoom = 4,
  bbox = NULL,
  recipe_name
)
}
\arguments{
\item{layer_names}{<\verb{named character vector}> All the layers that should be
part of the tileset. All the other arguments are named accordignly to what
is present in \code{layer_names}.}

\item{source}{<\verb{named character vector}> The tileset source of every layer.}

\item{minzoom}{<\verb{named numeric vector}> The minimum zoom levels at which
the features in the tileset should be visible.}

\item{maxzoom}{<\verb{named numeric vector}> The maximum zoom levels at which
the features in the tileset should be visible.}

\item{layer_size}{<\verb{named numeric vector}> Sets the maximum allowed size
(in KiB) of the tile layer}

\item{simp_zoom}{<\verb{named numeric vector}> At which zoom should the simplification
starts taking place.}

\item{simp_value}{<\verb{named numeric vector}> The simplification value. The normal
default simplification value is 4. The \code{simp_value} will take place for zoom
levels higher than the \code{simp_zoom}.}

\item{fallback_simp_zoom}{<\verb{named numeric vector}> The simplification value for
every zoom level below the \code{simp_zoom}}

\item{bbox}{<\verb{numeric vector}> Bounding box of the source, used for testing.  It
must contain an array of four numbers in this order: the minimum longitude,
minimum latitude, maximum longitude, and maximum latitude.}

\item{recipe_name}{<\code{character}> The ID of the new tile source. For a scale, usually
follows a prefix (mtl), the region (CMA), and the scale (CSD), e.g \code{mtl_CMA_CSD}.}
}
\value{
A JSON document containing configuration options that tell Mapbox Tiling
Service (MTS) how to turn tileset source data into vector tiles.
}
\description{
Create a recipe from scratch. Almost every argument takes a named vector as
for example the auto-zoom must present multiple features at different zoom levels.
So the \code{minzoom} must be a named vector, the name corresponding to one entry
in the \code{layer_names}. For more information on every argument, visit
https://docs.mapbox.com/mapbox-tiling-service/reference/
}
