% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilesets.R
\name{tileset_create_tileset}
\alias{tileset_create_tileset}
\title{Create a tileset}
\usage{
tileset_create_tileset(tileset, recipe, username, access_token)
}
\arguments{
\item{tileset}{<\code{character}> Name of the tileset that will be created}

\item{recipe}{<\code{character}> Tileset recipe previously created with
\code{\link[cc.buildr]{tileset_create_recipe}}}

\item{username}{<\code{character}> Mapbox account username.}

\item{access_token}{<\code{character}> Private access token to the Mapbox account.}
}
\value{
A success message if succeeds.
}
\description{
Create a tileset
}
