% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilesets.R
\name{tileset_delete_tileset}
\alias{tileset_delete_tileset}
\title{Delete a tileset}
\usage{
tileset_delete_tileset(id, username, access_token)
}
\arguments{
\item{id}{<\code{character}> ID of the tileset that needs to be deleted.}

\item{username}{<\code{character}> Mapbox account username.}

\item{access_token}{<\code{character}> Private access token to the Mapbox account.}
}
\value{
Returns nothing if succeeds.
}
\description{
Delete a tileset
}
