% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilesets.R
\name{tileset_delete_tileset_source}
\alias{tileset_delete_tileset_source}
\title{Delete tileset source}
\usage{
tileset_delete_tileset_source(id, username, access_token)
}
\arguments{
\item{id}{<\code{character}> ID of the tileset source that needs to be deleted.}

\item{username}{<\code{character}> Mapbox account username.}

\item{access_token}{<\code{character}> Private access token to the Mapbox account.}
}
\value{
A success message if succeeds.
}
\description{
Delete tileset source
}
