% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilesets.R
\name{tileset_labels}
\alias{tileset_labels}
\title{Create CSD labels tileset}
\usage{
tileset_labels(CSD_table, crs, prefix, username, access_token)
}
\arguments{
\item{CSD_table}{<\verb{sf data.frame}> The data.frame containing the most filled
set of CSDs. Their labels will be the \code{name} column, and the \code{population}
column will be used to weight which label to display first.}

\item{crs}{<\code{numeric}> EPSG coordinate reference system to be assigned, e.g.
\code{32617} for Toronto.}

\item{prefix}{<\code{character}> Prefix attached to every tile source and
created and published tileset. Should correspond to the Curbcut city, e.g. \code{mtl}.}

\item{username}{<\code{character}> Mapbox account username.}

\item{access_token}{<\code{character}> Private access token to the Mapbox account.}
}
\value{
Returns nothing if succeeds. Tilesets are created and published and
ready to be used.
}
\description{
Create CSD labels tileset
}
