% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilesets.R
\name{tileset_list_tile_sources}
\alias{tileset_list_tile_sources}
\title{List all tile sources in a Mapbox account}
\usage{
tileset_list_tile_sources(username, access_token)
}
\arguments{
\item{username}{<\code{character}> Mapbox account username.}

\item{access_token}{<\code{character}> Private access token to the Mapbox account.}
}
\value{
A data.frame listing all the tile sources in the account.
}
\description{
List all tile sources in a Mapbox account
}
