% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilesets.R
\name{tileset_publish_tileset}
\alias{tileset_publish_tileset}
\title{Publish a previously created tileset}
\usage{
tileset_publish_tileset(tileset, username, access_token)
}
\arguments{
\item{tileset}{<\code{character}> Name of the previously created tileset (with
\code{\link[cc.buildr]{tileset_create_tileset}}) which will be published.}

\item{username}{<\code{character}> Mapbox account username.}

\item{access_token}{<\code{character}> Private access token to the Mapbox account.}
}
\value{
A success message if succeeds.
}
\description{
Publish a previously created tileset
}
