% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilesets.R
\name{tileset_streets}
\alias{tileset_streets}
\title{Create and publish street and park tilesets}
\usage{
tileset_streets(
  master_polygon,
  street,
  crs,
  inst_prefix,
  username,
  access_token
)
}
\arguments{
\item{master_polygon}{<\verb{sf data.frame}> Unioned multipolygon of all the
geometries for which data has been gathered.}

\item{street}{<\verb{sf data.frame}> All the streets in the zone under study.}

\item{crs}{<\code{numeric}> EPSG coordinate reference system to be assigned, e.g.
\code{32617} for Toronto.}

\item{inst_prefix}{<\code{character}> Prefix attached to every tile source and
created and published tileset. Should correspond to the Curbcut city, e.g. \code{mtl}.}

\item{username}{<\code{character}> Mapbox account username.}

\item{access_token}{<\code{character}> Private access token to the Mapbox account.}
}
\value{
Returns nothing if succeeds. Tilesets are created and published and
ready to be used.
}
\description{
Create and publish street and park tilesets
}
