% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilesets.R
\name{tileset_upload_all}
\alias{tileset_upload_all}
\title{Do it all tileset function}
\usage{
tileset_upload_all(all_scales, map_zoom_levels, prefix, username, access_token)
}
\arguments{
\item{all_scales}{<\verb{named list}> A named list of sf data.frame
containing all scales listed with their regions, normally
\code{scales_variables_modules$scales}.}

\item{map_zoom_levels}{<\verb{named list}> The previously created zoom levels
using \code{\link[cc.buildr]{map_zoom_levels_create_all}} and
\code{\link[cc.buildr]{map_zoom_levels_create_custom}}.}

\item{prefix}{<\code{character}> Prefix attached to every tile source and
created and published tileset. Should correspond to the Curbcut city, e.g. \code{mtl}.}

\item{username}{<\code{character}> Mapbox account username.}

\item{access_token}{<\code{character}> Private access token to the Mapbox account.}
}
\value{
Returns nothing if succeeds. Tilesets are created and published and
ready to be used.
}
\description{
Constructs the combinations with the default auto-zoom, creates the tile
sources, the recipes, the creation and publishing of tilesets. For the auto-zoom
that will get attached for each region, only the first scale along with \code{CT},
\code{DA} and \code{building} will be part of the default region auto-zoom.
}
