% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilesets.R
\name{tileset_upload_custom_auto_zoom}
\alias{tileset_upload_custom_auto_zoom}
\title{TKTK NOT EXPORT, NEEDS REWORK Upload a custom auto_zoom}
\usage{
tileset_upload_custom_auto_zoom(
  smaller_limit_scale,
  auto_zoom_suffix,
  all_scales,
  prefix,
  username,
  access_token
)
}
\arguments{
\item{smaller_limit_scale}{<\code{character}> The scale which is the smaller limit
of the new auto-zoom, e.g. \code{CT} for an auto-zoom that stops at census tracts.}

\item{auto_zoom_suffix}{<\code{character}> To have a unique identifier to the new
auto-zoom tileset. If the smaller limit of the new auto-zoom is \code{CT}, we
could choose \code{max_CT}.}

\item{all_scales}{<\verb{named list}> A named list of sf data.frame
containing all scales listed with their regions, normally
\code{scales_variables_modules$scales}. The function will filter in all the scales
containing the \code{smaller_limit} scale.}

\item{prefix}{<\code{character}> Prefix attached to every tile source and
created and published tileset. Should correspond to the Curbcut city, e.g. \code{mtl}.}

\item{username}{<\code{character}> Mapbox account username.}

\item{access_token}{<\code{character}> Private access token to the Mapbox account.}
}
\value{
Returns nothing if succeeds. Tilesets are created and published and
ready to be used.
}
\description{
In cases where a new auto zoom needs to be created. While the normal auto-zoom
would usually be: CSD - CT - DA - building, for some modules missing data,
we might want to stop the auto zoom at CT.
}
