% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilesets.R
\name{tileset_upload_ndvi}
\alias{tileset_upload_ndvi}
\title{Upload NDVI Tilesets}
\usage{
tileset_upload_ndvi(
  grids_dir = "dev/data/built/",
  map_zoom_levels,
  max_zoom = list(grd600 = 11, grd300 = 13, grd120 = 14, grd60 = 15, grd30 = 16),
  regions,
  prefix,
  username,
  access_token,
  ndvi_delta_breaks = c(-0.5, -0.1, -0.02, 0.02, 0.1, 0.5),
  crs
)
}
\arguments{
\item{grids_dir}{<\code{character}> Directory containing grid files. Default is "dev/data/built/".}

\item{map_zoom_levels}{<\code{list}> List of zoom levels for different grid scales.}

\item{max_zoom}{<\code{list}> Maximum zoom levels for different grid scales.
Default values for scales grd600, grd300, grd120, grd60, and grd30.}

\item{regions}{<\code{data.frame} or \code{sf} object> Spatial data for different regions.}

\item{prefix}{<\code{character}> Prefix attached to every tile source and
created and published tileset. Should correspond to the Curbcut city, e.g. \code{mtl}.}

\item{username}{<\code{character}> Mapbox account username.}

\item{access_token}{<\code{character}> Private access token to the Mapbox account.}

\item{ndvi_delta_breaks}{<\verb{numeric vector}> Break points for categorizing NDVI deltas.
Default is c(-0.5, -0.1, -0.02, 0.02, 0.1, 0.5).}

\item{crs}{<\code{numeric}> EPSG coordinate reference system to be assigned, e.g.
\code{32617} for Toronto.}
}
\value{
Invisibly returns NULL. The function's primary purpose is the side-effects
of uploading, creating, and publishing tilesets.
}
\description{
This function uploads NDVI tilesets for different grid scales and regions.
It handles the deletion of old tilesets, creation of new ones, and their publishing.
The function is designed to work with specific data structures and assumes
the presence of NDVI year-specific data.
}
