% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_matrice.R
\name{tt_create_routing}
\alias{tt_create_routing}
\title{Download OSM and GTFS data, and build the network}
\usage{
tt_create_routing(master_polygon, routing_folder)
}
\arguments{
\item{master_polygon}{<\code{sfc_MULTIPOLYGON}>. Unioned multipolygon of all the
geometries for which census data must be gathered.}

\item{routing_folder}{<\code{character}> A folder where the routing files will be
downloaded.}
}
\value{
Download GTFS and OSM data in the \code{routing_folder} and the network.dat file
is created (Combine data inputs (gtfs, osm) in a directory to build a
multimodal transport network used for routing in R5) using \code{r5r}.
}
\description{
Download OSM and GTFS data, and build the network
}
