% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_matrice.R
\name{tt_get_db}
\alias{tt_get_db}
\title{Get dissemination block centroids}
\usage{
tt_get_db(
  master_polygon,
  region,
  street,
  census_dataset = cc.buildr::current_census,
  routing_folder,
  crs
)
}
\arguments{
\item{master_polygon}{<\code{sfc_MULTIPOLYGON}>. Unioned multipolygon of all the
geometries for which census data must be gathered.}

\item{region}{<\verb{named list}> A named list of census regions to retrieve.
Names must be valid census aggregation levels. Preferably a whole province
to make sure all geometries present in the <\code{master_polygon}> is added.
e.g. \code{list(PR = 24)} for Montreal. Preferably use the output of
cc.buildr::create_master_polygon()$province_cancensus_code.}

\item{street}{<\verb{sf data.frame}> All the streets segments in the study region.}

\item{census_dataset}{<\code{character}> The dataset to query for available
regions, e.g. \code{"CA16"}.}

\item{routing_folder}{<\code{character}> A folder where the routing files were
downloaded.}

\item{crs}{<\code{numeric}> EPSG coordinate reference system to be assigned, e.g.
\code{32618} for Montreal.}
}
\value{
Returns a point sf data.frame where every point attached to
dissemination blocks are on the street network the closest to the centroid.
}
\description{
Get dissemination block centroids
}
