% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data_files.R
\name{unload_scales}
\alias{unload_scales}
\title{Unload specified scales from the scales_variables_modules}
\usage{
unload_scales(scales, unload)
}
\arguments{
\item{scales}{<\code{list}> A list containing scales and
their associated variables and modules.}

\item{unload}{<\verb{character vector}> A character vector of scale names to
be unloaded from the list.}
}
\value{
<\code{list}> Returns the modified list of scales variables modules,
excluding the scales specified in \code{unload}.
}
\description{
This function removes specified scale variables from a list of scales.
It is useful for managing memory if a scale is not to be re-used.
}
