% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{update_task_URI}
\alias{update_task_URI}
\title{Update ECS task definition with a new image URI and deploy it}
\usage{
update_task_URI(
  app_name,
  new_URI,
  aws_cli_profile = Sys.getenv("AWS_ADMIN_PROFILE")
)
}
\arguments{
\item{app_name}{<\code{character}> The application name in ECS.}

\item{new_URI}{<\code{character}> New image URI to be used in the task definition.}

\item{aws_cli_profile}{<\code{character}> The AWS IAM profile which should be used
to upload the container to ECR. Must have the right permissions. Must have
been previously set through the AWS CLI using \verb{aws configure --profile X}.}
}
\value{
<\code{NULL}> This function is called for its side effects and does not return a value.
}
\description{
This function modifies the AWS ECS task definition to use a new image URI
and then updates the ECS service to deploy the new task definition. It assumes the
task names follow the pattern 'cc-x' instead of 'curbcut-x' and is designed for use
when a new ECR image is pushed and a force deployment is desired.
}
