% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_dictionaries.R
\name{verify_dictionaries}
\alias{verify_dictionaries}
\title{Verify the geos and the scales dictionary}
\usage{
verify_dictionaries(all_tables, geos_dictionary, scales_dictionary)
}
\arguments{
\item{all_tables}{<\verb{named list}> The name of the named list \code{all_tables}
is used to retrieve the shapefiles in the supplied folder.}

\item{geos_dictionary}{<\code{data.frame}> The geos dictionary built using
\code{\link[susbuildr]{geos_dictionary}}}

\item{scales_dictionary}{<\code{data.frame}> The scales dictionary built using
\code{\link[susbuildr]{build_census_scales}} and
\code{\link[susbuildr]{additional_scale}}}
}
\value{
Returns nothing if the test passes, or errors if it doesn't.
}
\description{
The use of \code{verify_dictionaries()} is to verify any missing entries in both
dictionaries.
}
