% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BC_page_zoning.R
\name{zoning_page}
\alias{zoning_page}
\title{BC Zoning Page}
\usage{
zoning_page(
  scales_variables_modules,
  base_polygons,
  username,
  access_token,
  crs,
  overwrite,
  overwrite_mb,
  tileset_prefix,
  region_DA_IDs,
  CMA_name_in_coding_xlsx
)
}
\arguments{
\item{scales_variables_modules}{<\code{list}> A list containing scales, variables,
and modules data.}

\item{base_polygons}{<\verb{sf object}> An sf object containing base polygons
for spatial filtering.}

\item{username}{<\code{character}> Mapbox username.}

\item{access_token}{<\code{character}> Mapbox access token.}

\item{crs}{<\code{numeric}> Coordinate Reference System code for spatial transformations.}

\item{overwrite}{<\code{logical}> Flag to indicate if existing files should be overwritten.}

\item{overwrite_mb}{<\code{logical}> Flag to indicate if existing Mapbox data
should be overwritten.}

\item{tileset_prefix}{<\code{character}> Prefix for naming the Mapbox tileset.}

\item{region_DA_IDs}{<\code{vector}> Vector of DA (Dissemination Area) IDs
for extracting frequent bus stops.}

\item{CMA_name_in_coding_xlsx}{<\code{character}> In the coding document, what is
the \code{CMA} corresponding to this region?}
}
\value{
<\code{list}> A list containing updated scales, variables, and modules.
}
\description{
This function processes zoning data and updates the map tileset. It filters and
transforms zoning data, calculates various zoning metrics, and uploads the
updated data to Mapbox for visualization. The function also updates the modules
table with zoning information.
}
