% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_table.R
\name{add_module}
\alias{add_module}
\title{Add a new module to the modules table}
\usage{
add_module(modules, id, title, metadata, dataset_info, geos = NULL)
}
\arguments{
\item{modules}{<\code{data.frame}> The \code{modules} data.frame to which add a
new row.}

\item{id}{<\code{character}> The module id, e.g. \code{canale}}

\item{title}{<\code{character}> The module title which will be used to navigate
to that module from the navigation bar, e.g. \verb{Active living potential}}

\item{metadata}{<\code{logical}> Indicate if metadata is available from this module,
if there should be further data info when the \verb{Export data} popup is clicked
(information regarding if the data has been interpolated, source, etc.).}

\item{dataset_info}{<\code{character}> HTML text with further data information
from the module.}

\item{geos}{<\verb{character vector}> List of all the geos the module should be
able to showé}
}
\value{
The same \code{modules} data.frame fed, with the added row.
}
\description{
Add a new module to the modules table
}
