% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_table.R
\name{add_module}
\alias{add_module}
\title{Add a new module to the modules table}
\usage{
add_module(
  modules,
  id,
  nav_title,
  title_text_title,
  title_text_main,
  title_text_extra,
  metadata,
  dataset_info,
  geos = NULL
)
}
\arguments{
\item{modules}{<\code{data.frame}> The \code{modules} data.frame to which add a
new row.}

\item{id}{<\code{character}> The module id, e.g. \code{canale}}

\item{nav_title}{<\code{character}> The module title which will be used to navigate
to that module from the navigation bar, e.g. \verb{Active living potential}}

\item{title_text_title}{<\code{character}> This title text title will be used
once the user is in the module, to describe what they are navigating,
e.g. \verb{Active living potential: the CanALE index}.
to that module from the navigation bar, e.g. \verb{Active living potential}}

\item{title_text_main}{<\code{character}> The title text main paragraph is the main
text used to describe the module, on the left-side panel.}

\item{title_text_extra}{<\code{character}> Further information and resources the user
can navigate if they press on the 'Learn more' button in the left-side panel.}

\item{metadata}{<\code{logical}> Indicate if metadata is available from this module,
if there should be further data info when the \verb{Export data} popup is clicked
(information regarding if the data has been interpolated, source, etc.).}

\item{dataset_info}{<\code{character}> HTML text with further data information
from the module.}

\item{geos}{<\verb{character vector}> List of all the geos the module should be
able to show.}
}
\value{
The same \code{modules} data.frame fed, with the added row.
}
\description{
Add a new module to the modules table
}
