% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{add_q5}
\alias{add_q5}
\title{Append q5 column to a dataframe}
\usage{
add_q5(df, breaks, time_regex = "_\\\\d{4}$")
}
\arguments{
\item{df}{<\code{data.frame}> Contains all columns in \code{vars}.}

\item{breaks}{<\code{data.frame}> A data.frame containing all the q5 breaks. The
output of \code{\link[cc.buildr]{get_breaks_q5}}.}

\item{time_regex}{<\code{character}> Regular expression which corresponds to
a timeframe, placed at the end of the \code{vars} vector. e.g. \verb{_\\\\d\{4\}$} for
years.}
}
\value{
Returns the same data.frame as df with q5 columns appended.
}
\description{
Append q5 column to a dataframe
}
