% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_table.R
\name{add_var_right}
\alias{add_var_right}
\title{Add right variables to Modules}
\usage{
add_var_right(svm)
}
\arguments{
\item{svm}{<\code{list}> scales_variables_modules list object.}
}
\value{
<\code{list}> The modified \code{svm} object with \code{var_right} added to each
module in the \code{modules} dataframe.
}
\description{
This function iterates over the modules in a \code{scales_variables_modules} object,
identifying and adding the appropriate right variables based on the classification
of left-side variables. It ensures that all left-side variables within a module
have the same classification before proceeding.
}
