% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables_table.R
\name{add_variable}
\alias{add_variable}
\title{Add a new variable to the variables table}
\usage{
add_variable(
  variables,
  var_code,
  type,
  var_title,
  var_short = as.character(var_title),
  explanation,
  group_name = NA_character_,
  group_diff = list(),
  theme,
  private,
  dates,
  scales,
  breaks_q3,
  breaks_q5,
  source,
  interpolated
)
}
\arguments{
\item{variables}{<\code{data.frame}> The \code{variables} data.frame to which add a
new row.}

\item{var_code}{<\code{character}> The code used to refer to the variables, e.g.
\code{canale}}

\item{type}{<\code{character}> The variable type. One of \code{"ind"}, \code{"pct"},
\code{"avg"}, \code{"median"}, \code{"per1k"}, \code{"sqkm"} or \code{"count"}.}

\item{var_title}{<\code{character}> The variable title}

\item{var_short}{<\code{character}> A short variable title used in graphs or where
space is limited. Preferably ~ <12 characters.}

\item{explanation}{<\code{character}> Variable explanation, e.g. the percentage of
private dwellings occupied by tenants}

\item{group_name}{<\code{character}> The name of the larger group to which the
variable belongs. e.g. for the variable accessibility to public schools by bike,
the group_name would be \code{"Accessibility to schools"}}

\item{group_diff}{<\verb{named list}> A named list for when the variable is part
of a greater group. e.g. accessibility to public schools by bike, the bigger group
is \verb{Accessibility to schools}, and bike is a group differentiation.
e.g. The \code{list("Mode of transport" = "By bike", "Public/Prviate" = "Public")}.
The list can contain multiple named vectors, multiple group differentiation.}

\item{theme}{<\code{character}> The theme to which the variable belongs, e.g. "Housing",
"Urban life", ...}

\item{private}{<\code{logical}> If we have permissions to make the variable available
for public download.}

\item{dates}{<\verb{character vector}> A vector of dates for which the data is available.}

\item{scales}{<\code{data.frame}> All the scales at which the data is available
arranged in a data.frame of two columns: \code{geo} and \code{scale}}

\item{breaks_q3}{<\code{data.frame}> A data.frame with with information regarding
scales, date, rank, breaks. The lst outputs of \code{\link[susbuildr]{calculate_breaks}}}

\item{breaks_q5}{<\code{data.frame}> A data.frame with with information regarding
scales, date, rank, breaks. The lst outputs of \code{\link[susbuildr]{calculate_breaks}}}

\item{source}{<\code{character}> The source where the data comes from, e.g.
"McGill Geo-Social Determinants of Health Research Group"}

\item{interpolated}{<\code{data.frame}> A data.frame indicating from which scale
the geo/scale comination has been interpolated. The non-interpolated data
is populated with \code{"FALSE"}}
}
\value{
The same \code{variables} data.frame fed, with the added row.
}
\description{
Add a new variable to the variables table
}
