% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables_table.R
\name{add_variable}
\alias{add_variable}
\title{Add a new variable to the variables table}
\usage{
add_variable(
  variables,
  var_code,
  type,
  var_title,
  var_short = as.character(var_title),
  explanation,
  exp_q5,
  parent_vec,
  group_name = NA_character_,
  group_diff = list(),
  theme,
  private,
  pe_include = TRUE,
  dates,
  avail_df,
  breaks_q3,
  breaks_q5,
  region_values = NULL,
  source,
  interpolated,
  rankings_chr = c("exceptionally low", "unusually low", "just about average",
    "unusually high", "exceptionally high"),
  var_measurement = data.frame(df = avail_df, measurement = rep("scalar",
    length(avail_df)))
)
}
\arguments{
\item{variables}{<\code{data.frame}> The \code{variables} data.frame to which add a
new row.}

\item{var_code}{<\code{character}> The code used to refer to the variables, e.g.
\code{canale}}

\item{type}{<\code{character}> The variable type. One of \code{"ind"}, \code{"pct"},
\code{"avg"}, \code{"median"}, \code{"per1k"}, \code{"sqkm"} or \code{"count"}.}

\item{var_title}{<\code{character}> The variable title}

\item{var_short}{<\code{character}> A short variable title used in graphs or where
space is limited. Preferably ~ <12 characters.}

\item{explanation}{<\code{character}> Variable explanation, e.g. the percentage of
private dwellings occupied by tenants}

\item{exp_q5}{<\code{character}> String used for the explore panel explaining the
variable. Depends on the \code{type}. The rules are:
\itemize{
\item{"pct"}{starts with a verb and follows the absolute count of the parent
string. the string will read e.g. '50 households (3\%) are tenants' where
\verb{are tenants} is the definition value.}
\item{"dollar"}{starts with a subject and ends with a verb. It's assumed it
will be followed by a dollar number. e.g. 'the average rent is 800$' where
\verb{the average rent is} is the definition value.}
\item{"ind"}{starts with a verb and follow the absolute count of the parent
string. Uses a place holder written \verb{_X_} which would translates to, e.g. 'medium to high'.
example: '50 households are living in areas with low potential for active living'
where the definition would be: \verb{are living in areas with _X_ potential for active living}}
}}

\item{parent_vec}{<\code{character}> Parent vector of the variable. Used for
the explore panel. Must be another entry in the variable table. E.g. for
Tenant households (\%), the parent variable would be the number of
private households (denominator of the percentage) : \code{private_households}.}

\item{group_name}{<\code{character}> The name of the larger group to which the
variable belongs. e.g. for the variable accessibility to public schools by bike,
the group_name would be \code{"Accessibility to schools"}}

\item{group_diff}{<\verb{named list}> A named list for when the variable is part
of a greater group. e.g. accessibility to public schools by bike, the bigger group
is \verb{Accessibility to schools}, and bike is a group differentiation.
e.g. The \code{list("Mode of transport" = "By bike", "Public/Prviate" = "Public")}.
The list can contain multiple named vectors, multiple group differentiation.}

\item{theme}{<\code{character}> The theme to which the variable belongs, e.g. "Housing",
"Urban life", ...}

\item{private}{<\code{logical}> If we have permissions to make the variable available
for public download.}

\item{pe_include}{<\code{logical}> Should this variable be included in the place
explorer? Defaults to \code{TRUE}.}

\item{dates}{<\verb{character vector}> A vector of dates for which the data is available.}

\item{avail_df}{<\code{list}> All the combinations of region and scales
at which the data is available, e.g. \code{c("CMA_CSD", "CMA_CT", ...)}}

\item{breaks_q3}{<\code{data.frame}> A data.frame with with information regarding
scales, date, rank, breaks. The last outputs of
\code{\link[cc.buildr]{calculate_breaks}}}

\item{breaks_q5}{<\code{data.frame}> A data.frame with with information regarding
scales, date, rank, breaks. The last outputs of
\code{\link[cc.buildr]{calculate_breaks}}}

\item{region_values}{<\code{data.frame}> A data.frame with information regarding
the values of the overall region. Must include the region, and potentially
the year, val and count columns, depending on the variable type. Can be
created using \code{\link{variables_get_region_vals}}.}

\item{source}{<\code{character}> The source where the data comes from, e.g.
"McGill Geo-Social Determinants of Health Research Group"}

\item{interpolated}{<\code{data.frame}> A data.frame indicating from which scale
the geo/scale comination has been interpolated. The non-interpolated data
is populated with \code{"FALSE"}}

\item{rankings_chr}{<\verb{character vector}> Vector of character that will be used
in the explore text to inform how a location ranks within the region.
The attach character will be its \code{q5} break. \code{ranking_chr} must be in order of
lower to higher. \verb{The variable score is 90\% which is unusually low for Montreal}.
Defaults to \code{c("exceptionally low", "unusually low", "just about average", "unusually high", "exceptionally high")}}

\item{var_measurement}{<\verb{character list}> Data.frame where every row is an available
df (\code{aval_df}) and the second column is the measurement type. These will impact
both the explore panel text and the explore panel graph. Options are:
\itemize{
\item{"scalar"}{The default. If \code{var_measurement} is not supplied, the row will
be automatically filled with this measurement. Represents numeric data with both
meaningful order and consistent distance between values. Example: age, temperature,
height, weight.}
\item{"ordinal"}{Represents rank or order within a dataset. It has meaningful
order or sequence, but the distance between values is not consistent. It is
the one used for the climate risk page at the grid scale where
1 = Insignificant vulnerability, 2 = Minore, 3 = Moderate, ...
Other example: survey responses with options like 'Strongly disagree', 'Disagree', ...}
\item{"nominal"}{Represents categories without inherent order. No meaningful
order or ranking is possible. Qualitative variables. Example: animal types,
such as 'Mammal', 'Bird', 'Reptile', 'Fish'.}
}}
}
\value{
The same \code{variables} data.frame fed, with the added row.
}
\description{
Add a new variable to the variables table
}
