% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_census_data.R
\name{ba_census_data}
\alias{ba_census_data}
\title{Build and append census data}
\usage{
ba_census_data(
  scales_variables_modules,
  region_DA_IDs,
  census_vectors = cc.data::census_vectors,
  census_years = cc.data::census_years,
  skip_scale_interpolation = NULL,
  scales_sequences,
  crs,
  housing_module = TRUE
)
}
\arguments{
\item{scales_variables_modules}{<\verb{named list}> A list of length three.
The first is all the scales, the second is the variables table, and the
third is the modules table.}

\item{region_DA_IDs}{<\verb{character vector}> All the current census'
DA IDs present in the region. Only those will be extracted from the database
to do interpolation.}

\item{census_vectors}{<\verb{character vector}> Data variables that should be added
to the scales. By default, all: \code{\link[cc.data]{census_vectors}}. Look
at the \code{\link[cc.data]{census_vectors_table}} to view all
variables explained.}

\item{census_years}{<\verb{character vector}> Years for which the census data
should be added to the scales. Defaults to \code{\link[cc.data]{census_years}}}

\item{skip_scale_interpolation}{<\verb{character vector}> Scales for which census
data should not be interpolated (e.g. very small scales like 25m grid cells.).
In those cases, census data won't be interpolated and appended. Defaults to
NULL to interpolate to everything. The interpolation function will detect
scales smaller than DAs and automatically won't do interpolation for them.}

\item{scales_sequences}{<\code{list}> A list of scales sequences representing the
hierarchical ordering of scales on an auto-zoom.}

\item{crs}{<\code{numeric}> EPSG coordinate reference system to be assigned, e.g.
\code{32617} for Toronto.}

\item{housing_module}{<\code{logical}> Should a housing module be added to
the list of modules.}
}
\value{
A list of length 3, similar to the one fed to
\code{scales_variables_modules} with census variable added, their addition
in the variables table and the module table.
}
\description{
Build and append census data
}
