% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_census_data.R
\name{ba_census_data}
\alias{ba_census_data}
\title{Build and append census data}
\usage{
ba_census_data(
  scales_variables_modules,
  region_DA_IDs,
  census_vectors = cc.data::census_vectors,
  census_years = cc.data::census_years,
  scales_to_interpolate = {
    
    names(scales_variables_modules$scales)[!names(scales_variables_modules$scales) \%in\%
    c("building", "street", "DB", "grd30", "grd60", "grd120", "grd300")]
 },
  scales_sequences,
  crs,
  DB_table,
  housing_module = TRUE,
  age_module = TRUE,
  overwrite = FALSE,
  inst_prefix
)
}
\arguments{
\item{scales_variables_modules}{<\verb{named list}> A list of length three.
The first is all the scales, the second is the variables table, and the
third is the modules table.}

\item{region_DA_IDs}{<\verb{character vector}> All the current census'
DA IDs present in the region. Only those will be extracted from the database
to do interpolation.}

\item{census_vectors}{<\verb{character vector}> Data variables that should be added
to the scales. By default, all: \code{\link[cc.data]{census_vectors}}. Look
at the \code{\link[cc.data]{census_vectors_table}} to view all
variables explained.}

\item{census_years}{<\verb{character vector}> Years for which the census data
should be added to the scales. Defaults to \code{\link[cc.data]{census_years}}}

\item{scales_to_interpolate}{<\verb{character vector}> Scales for which census
data should be interpolated (e.g. very small scales like 25m grid cells should be excluded.).
Defaults to all the scales except building and street.}

\item{scales_sequences}{<\code{list}> A list of scales sequences representing the
hierarchical ordering of scales on an auto-zoom.}

\item{crs}{<\code{numeric}> EPSG coordinate reference system to be assigned, e.g.
\code{32617} for Toronto.}

\item{DB_table}{<\verb{named list}> Named list of all the years for which DBs existed.
Every dataframe in the list must have the DB ID, DA ID, population and dwellings
count. It is created using \code{\link[cc.data]{DB_get}}.}

\item{housing_module}{<\code{logical}> Should a housing module be added to
the list of modules.}

\item{age_module}{<\code{logical}> Should an age module be added to
the list of modules.}

\item{overwrite}{<\code{logical}> Should the data already processed and stored be
overwriten?}

\item{inst_prefix}{<\code{character}> The prefix of the instance, e.g. \code{'mtl'} which
is the database schema in which the data is saved.}
}
\value{
A list of length 3, similar to the one fed to
\code{scales_variables_modules} with census variable added, their addition
in the variables table and the module table.
}
\description{
Build and append census data
}
