% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_census_data.R
\name{build_census_data}
\alias{build_census_data}
\title{Build census data at all possible scales}
\usage{
build_census_data(
  scales_consolidated,
  region_DA_IDs,
  census_vectors,
  census_years,
  crs
)
}
\arguments{
\item{scales_consolidated}{A named list of sf data.frame
containing all scales listed with their regions. The output of
\code{\link[cc.buildr]{consolidate_scales}}.}

\item{region_DA_IDs}{<\verb{character vector}> All the current census'
DA IDs present in the region. Only those will be extracted from the database
to do interpolation.}

\item{census_vectors}{<\verb{character vector}> Data variables that should be added
to the scales. By default, all: \code{\link[cc.data]{census_vectors}}. Look
at the \code{\link[cc.data]{census_vectors_table}} to view all
variables explained.}

\item{census_years}{<\verb{character vector}> Years for which the census data
should be added to the scales. Defaults to \code{\link[cc.data]{census_years}}}

\item{crs}{<\code{numeric}> EPSG coordinate reference system to be assigned, e.g.
\code{32617} for Toronto.}
}
\value{
Returns a list of length 3. The first is the possible scales for which
census data can be added in scales_consolidated. The second is a
data.frame of scales reference, to know for which scales the data is
available. It can directly be fed to the scales argument of add_variable.
The third is a data.frame of interpolation reference, to know for which
scales the data has been interpolated. It can directly be fed to the
interpolated argument of add_variable.
}
\description{
Build census data at all possible scales
}
