% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_census_scales.R
\name{build_census_scales}
\alias{build_census_scales}
\title{Build census scales}
\usage{
build_census_scales(
  master_polygon,
  census_dataset = susbuildr::current_census,
  regions,
  levels = c("CSD", "CT", "DA"),
  crs,
  fill_CTs_with_CSDs = TRUE,
  override_name_2 = list(CSD = "City")
)
}
\arguments{
\item{master_polygon}{<\code{sfc_MULTIPOLYGON}>. Unioned multipolygon of all the
geometries for which census data must be gathered.}

\item{census_dataset}{<\code{character}> The dataset to query for available
regions, e.g. \code{"CA16"}. Defaults to \code{\link[susbuildr]{current_census}}.}

\item{regions}{<\verb{named list}> A named list of census regions to retrieve.
Names must be valid census aggregation levels. Preferably a whole province
to make sure all geometries present in the <\code{master_polygon}> is added.
e.g. \code{list(PR = 24)} for Montreal.}

\item{levels}{<\code{character}> The census aggregation levels to retrieve.
Can be \code{"CMA"}, \code{"CD"}, \code{"CT"}, \code{"DA"}, \code{"DB"}, ...}

\item{crs}{<\code{numeric}> EPSG coordinate reference system to be assigned, e.g.
\code{32618} for Montreal.}

\item{fill_CTs_with_CSDs}{<\code{logical}> If the master polygon is bigger than a
CMA, should what is outside the CMA display the CSDs on the CT scale?}

\item{override_name_2}{<\verb{named list}> Override name_2. Defaults to name_2 for
CSDs being "City" with \code{list(CSD = "City")}, e.g. the display on Sus
for CSDs would then be e.g. \verb{City of Laval}. For scales below CSD, the name
of the CSD in which the zone is gets the CSD name as their name_2. The
default display for a CT would then be e.g. \verb{Census tract 4620633.00 (Laval)}.}
}
\value{
A list of sf dataframes of census scales filtered by the master polygon,
with the option of one CSD subdivided.
}
\description{
Build census scales
}
