% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{calculate_breaks}
\alias{calculate_breaks}
\title{Calculate all breaks when all_scales is a list of regions and scales}
\usage{
calculate_breaks(all_scales, vars, time_regex = "\\\\d{4}")
}
\arguments{
\item{all_scales}{<\code{named_list}> A named list of scales. The first level is
the geo, and the second is the scales. They must contain all columns of \code{vars}}

\item{vars}{<\verb{vector of character}> Contains all variable names from which
to add q3s. Must fit with a name in \code{df}.}

\item{time_regex}{<\code{character}> Regular expression which corresponds to
a timeframe, placed at the end of the \code{vars} vector. e.g. \verb{\\\\d\{4\}} for
years. If the variable does not have a timeframe, enter an empty string (\code{""})}
}
\value{
Returns a list of length 4. The first is the same data.frame as df
with q3 and q5 columns appended. The second is the q3 breaks table, and the third
is the q5 breaks table. The fourth is vectors of characters of all dates at
which variables are available for.
}
\description{
Calculate all breaks when all_scales is a list of regions and scales
}
