% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_master_polygon.R
\name{create_master_polygon}
\alias{create_master_polygon}
\title{Create master polygon}
\usage{
create_master_polygon(all_tables, shapefiles_folder, crs)
}
\arguments{
\item{all_tables}{<\verb{named list}> The name of the named list \code{all_tables}
is used to retrieve the shapefiles in the supplied folder.}

\item{shapefiles_folder}{<\code{character}> The file path of where to find
the shapefiles of the geometries to read (names of all_table).}

\item{crs}{<\code{numeric}> EPSG coordinate reference system to be assigned, e.g.
\code{32618} for Montreal.}
}
\value{
The output is a named list of length 2: The master polygon
and all the individual geos present in \code{all_tables}.
}
\description{
Create master polygon
}
