% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_page_script.R
\name{create_page_script}
\alias{create_page_script}
\title{Create page script}
\usage{
create_page_script(source_file, R_folder = "R/", overwrite = FALSE)
}
\arguments{
\item{source_file}{<\code{character}> File where all inputs are like module id,
var_left, var_right, time, etc.}

\item{R_folder}{<\code{character}> Location where the new page script will be created.
Defaults to \code{"R/"}}

\item{overwrite}{<\code{logical}> If a file already exists, should it be overriden.
Defaults to \code{FALSE}. The file name will be made using the \code{id} object from
the \code{source_file}. \code{R/m_id.R}.}
}
\value{
Opens the new page script created.
}
\description{
Create page script
}
