% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyk.R
\name{dyk_uni}
\alias{dyk_uni}
\title{Create a Table of Univariate DYKs}
\usage{
dyk_uni(vars_dyk, svm)
}
\arguments{
\item{vars_dyk}{<\code{data.frame}> A data frame with columns \code{module},
\code{region}, \code{scale}, \code{var_left}, \code{var_right} and \code{date}, probably created with
\code{\link[cc.buildr]{dyk_prep}}.}

\item{svm}{<\code{list}> A list, usually \code{scales_variables_modules}, containing
the scales, modules, and variables tables.}
}
\value{
A data frame with nine columns (the columns present in \code{vars_dyk},
and then additionally \code{dyk_type}, \code{dyk_text} and \code{dyk_value}).
}
\description{
This function creates a table of univariate "Did you know" text strings.
}
