% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyk.R
\name{dyk_uni_highest_lowest}
\alias{dyk_uni_highest_lowest}
\title{Generate Highest/Lowest Value DYKs}
\usage{
dyk_uni_highest_lowest(var_left, region, scale, date, svm)
}
\arguments{
\item{var_left}{\if{html}{\out{<character>}} A string representing the name of the variable
for which the DYK should be calculated.}

\item{region}{<\code{character}> A string representing the name of the region
for which the DYK should be calculated.}

\item{scale}{<\code{character}> A string representing the name of the scale
for which the DYK should be calculated.}

\item{date}{<\code{character}> A string representing the name of the date
for which the DYK should be calculated.}

\item{svm}{<\code{list}> A list, usually \code{scales_variables_modules}, containing
the scales, modules, and variables tables.}
}
\value{
A data frame with two columns (\code{highest} and \code{lowest}), each of
which contains a character vector of DYK outputs.
}
\description{
This function creates "Did you know" text strings highlighting the highest
and lowest values from a combination of variable, region, scale and date.
The output is a data frame containing a \code{highest} and \code{lowest} column.
}
