% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geos_dictionary.R
\name{geos_dictionary}
\alias{geos_dictionary}
\title{Build the geos dictionary}
\usage{
geos_dictionary(all_tables, geo, name, to_compare)
}
\arguments{
\item{all_tables}{<\verb{named list}> Named list of geos and their scales within,
ordered in priority.}

\item{geo}{<\verb{vector of character}> A vector of characters used to identify
which large geometry the user is interested in, e.g.
\code{c("CMA", "city", "island")}.}

\item{name}{<\verb{named character vector}> Named with their corresponding geo value.
Same length as the vector fed in the \code{geo} argument. Used to name the geo
when the user will want to change which large geometry they are interested in. e.g.
\code{c("Metropolitan Area", "City of Montreal", "Island of Montreal")}.}

\item{to_compare}{<\verb{named character vector}> Named with their corresponding geo value.
Same length as the vector fed in the \code{geo} argument. The same length as
the vector fed in the \code{geo} argument. Used at the end of dynamically generated
text, when a particular zone is compared with the whole geography. 'x DA
has a higher value than y\% of DAs in the Montreal region'.
e.g. \code{c("in the Montreal region", "in the City of Montreal",
"on the island of Montreal")}}
}
\value{
Returns the same vectors fed arranged in a data.frame ordered in
priorty.
}
\description{
Build the geos dictionary
}
