% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heroku.R
\name{heroku_deploy}
\alias{heroku_deploy}
\title{Deploy an app in Heroku}
\usage{
heroku_deploy(app_name)
}
\arguments{
\item{app_name}{<\code{character}> Must correspond to an existing app on the
Heroku account that the user will login.}
}
\value{
Opens a terminal and disconnect from the current R session.
}
\description{
The function initiates the heroku login process and authenticates the
container. After successful authentication, it pushes the web application
to the specified Heroku account and finally, releases the web application.
}
