% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_zoom_levels.R
\name{map_zoom_levels_create_all}
\alias{map_zoom_levels_create_all}
\title{Create all basic map zoom levels}
\usage{
map_zoom_levels_create_all(
  all_tables,
  first_scale_zoom = 0,
  CT_zoom = 10.5,
  DA_zoom = 12.5,
  building_zoom = 15.5,
  building_or_street = "building"
)
}
\arguments{
\item{all_tables}{<\verb{named list}> A named list of character. The names are
the \code{region} e.g. CMA, island, city, ... and the vectors are all the
scales available in these geo e.g. CSD, CT, DA, building, ...}

\item{first_scale_zoom}{<\code{numeric}>¨On auto zoom, the zoom that should be
attributed to the first scale in a list of scales under a region.}

\item{CT_zoom}{<\code{numeric}>¨On auto zoom, the zoom that should be
attributed to the census tracts.}

\item{DA_zoom}{<\code{numeric}>¨On auto zoom, the zoom that should be
attributed to the dissemination ares.}

\item{building_zoom}{<\code{numeric}>¨On auto zoom, the zoom that should be
attributed to the buildings.}

\item{building_or_street}{<\code{character}>¨Which is the default smaller scale?
\code{building} or \code{street}.}
}
\value{
Returns a named list of all basic map zoom levels.
}
\description{
These zoom levels are used used to inform the zoom slider. The default
goes: first scale in the region, then \code{CT}, \code{DA} and \code{building} or \code{street}.
}
