% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_explorer.R
\name{placeex_main_card_rmd}
\alias{placeex_main_card_rmd}
\title{Pre-process all the possible Rmds}
\usage{
placeex_main_card_rmd(
  pe_main_card_data,
  regions_dictionary,
  scales_dictionary,
  lang = "en",
  tileset_prefix,
  mapbox_username = "curbcut",
  rev_geocode_from_localhost = TRUE,
  check_bslib_version = TRUE,
  overwrite = TRUE,
  skip_scales = c("building", "street", "grd30", "grd60", "grd120", "grd300", "grd600"),
  scales_sequences,
  full_data_path = sprintf("\%s/data/", getwd())
)
}
\arguments{
\item{pe_main_card_data}{<\code{list}> Data and dictionary necessary to knit the
rmds. The output of \code{\link[cc.buildr]{placeex_main_card_data}}.}

\item{regions_dictionary}{<\code{data.frame}> The regions dictionary built using
\code{\link[cc.buildr]{regions_dictionary}}. Will be used to filter out scales
for which data should not be calculated.}

\item{scales_dictionary}{<\code{data.frame}> The scales dictionary built using
\code{\link[cc.buildr]{build_census_scales}}}

\item{lang}{<\code{character}> If the Curbcut instance is bilingual, we can use
\code{c("en", "fr")} and HTML will be built for both languages. Defaults to \code{"en"}.}

\item{tileset_prefix}{<\code{character}> Prefix attached to every tile source and
created and published tileset. Should correspond to the Curbcut city, e.g. \code{mtl}.}

\item{mapbox_username}{<\code{character}> Mapbox account username.}

\item{rev_geocode_from_localhost}{<\code{character}> Use a local instance of nominatim
for reverse geocoding, using \code{\link[cc.data]{rev_geocode_localhost}}. If
TRUE, there must be a local instance of Nominatim runnin on localhost port 8080.
If set to FALSE, the reverse geolocation will be done using
\code{\link[cc.data]{rev_geocode_OSM}}, which uses a remote server (photon.komoot.io).}

\item{check_bslib_version}{<\code{logical}> The \code{bslib} library do not allow to
style individually each tab in a navigation bar. As we want to color each
tab depending on how much of an outlier the zone is in each theme, we have to
add \code{class} arguments to the navigation bar's list creation. This version can
be found at \code{devtools::install_github('bdbmax/bslib')}. Defaults to \code{TRUE.}
Set \code{FALSE} to continue with the version of \code{bslib} on your system.}

\item{overwrite}{<\code{logical}> Should the .html files be overwritten? Defaults
to \code{TRUE}.}

\item{skip_scales}{<\verb{character vector}> Scales for which place explorer documents
should not be created.}
}
\value{
Returns nothing if successful. All place explorer possibilities are
saved in the \code{out_folder}.
}
\description{
Pre-process all the possible Rmds
}
