% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reorder_columns}
\alias{reorder_columns}
\title{Reorder columns as they are expected in the workflow}
\usage{
reorder_columns(all_scales)
}
\arguments{
\item{all_scales}{<\verb{named list}> A named list of scales. The first level is
the geo, and the second is the scales.}
}
\value{
Returns the same list that is fed, with columns re-ordered.
}
\description{
Reorder columns as they are expected in the workflow
}
