% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data_files.R
\name{save_all_scales_qs}
\alias{save_all_scales_qs}
\title{Save all scale tables in QS format}
\usage{
save_all_scales_qs(data_folder = "data/", svm_data)
}
\arguments{
\item{data_folder}{<\code{character}> Where the \code{.qs} files should be
written to. Defaults to \verb{data/}.}

\item{svm_data}{<\verb{named list}> A named list of data.frame
containing all scales with a named list of their data. Usually
\code{scales_variables_modules$data}.}
}
\value{
An invisible NULL value.
}
\description{
This function saves all scale tables in QS format in the specified data folder.
It first drops the geometry of other scales and creates a list of tables for all
scales, then creates the folders for each region and scale, and finally saves the
tables in the database.
}
