% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data_files.R
\name{save_all_scales_qs}
\alias{save_all_scales_qs}
\title{Save all scale tables in QS format}
\usage{
save_all_scales_qs(data_folder = "data/", all_scales, variables)
}
\arguments{
\item{data_folder}{<\code{character}> Where the \code{.qs} files should be
written to. Defaults to \verb{data/}.}

\item{all_scales}{<\verb{named list}> A named list of sf data.frame
containing all scales listed with their regions, normally
\code{scales_variables_modules$scales}.}

\item{variables}{<\code{data.frame}> The \code{variables} data.frame, normally
\code{scales_variables_modules$variables}.}
}
\value{
An invisible NULL value.
}
\description{
This function saves all scale tables in QS format in the specified data folder.
It first drops the geometry of other scales and creates a list of tables for all
scales, then creates the folders for each region and scale, and finally saves the
tables in the database.
}
