% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data_files.R
\name{save_bslike_postgresql}
\alias{save_bslike_postgresql}
\title{Save df to PostgreSQL Database}
\usage{
save_bslike_postgresql(scale_chr, all_scales, inst_prefix)
}
\arguments{
\item{scale_chr}{<\code{character}> A character string specifying the scale key in
the \code{all_scales} list from which the data frame will be selected.}

\item{all_scales}{<\code{character}> A named list of data frames, where each key
represents a scale and each value is a data frame to be saved to the database
under the corresponding scale name.}

\item{inst_prefix}{<\code{character}> A string specifying the database schema, which is
the local prefix which is also the same prefix of tilesets. E.g. for Montreal, \code{mtl}.}
}
\value{
Invisibly returns \code{NULL}. The primary effect is the side effect of
writing the selected data frame to the PostgreSQL database.
}
\description{
This function saves a data frame to a PostgreSQL database table, selecting the
data frame from all_scales based on the scale name (scale_chr). The table
is written using the \code{db_write_prod} function, which handles the connection,
writing, and primary key setting.
}
