% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data_files.R
\name{save_buildings_sqlite}
\alias{save_buildings_sqlite}
\title{Save all buildings dataset in a SQLite}
\usage{
save_buildings_sqlite(scale_chr = "building", all_scales)
}
\arguments{
\item{scale_chr}{<\code{character}> The name of the scale: "building}

\item{all_scales}{<\verb{named list}> A named list of sf data.frame
containing all scales listed with their regions, normally
\code{scales_variables_modules$scales}.}
}
\value{
Returns an error or nothing if ran successfully. All existing \code{building}
data.frame in the fed \code{all_scales} are saved in the created \code{.sqlite}.
}
\description{
This function uses \code{\link{save_bslike_sqlite}}.
}
