% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stories.R
\name{stories_empty_table}
\alias{stories_empty_table}
\title{Create an empty data table for stories}
\usage{
stories_empty_table()
}
\value{
A tibble with columns \code{id}, \code{title}, \code{short_title},
\code{preview}, \code{themes}, \code{lon}, and \code{lat}.
}
\description{
Create an empty data table for stories
}
