% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stories.R
\name{stories_knit_all}
\alias{stories_knit_all}
\title{Knit All Stories}
\usage{
stories_knit_all(
  stories_location = "dev/Rmd/stories/",
  css_path = system.file("styles", "fonts.css", package = "curbcut"),
  output_dir = "www/stories/"
)
}
\arguments{
\item{stories_location}{<\code{character}> The path to the directory containing
the R Markdown story files. Default is "dev/Rmd/stories/".}

\item{css_path}{<\code{character}> A character string specifying the path to the
CSS file to be applied to the rendered HTML files. Default is "www/sus.css".}

\item{output_dir}{<\code{character}> A character string specifying the directory
where the rendered HTML files will be saved. Default is "www/stories/".}
}
\value{
A list of the knitted HTML files, with each item corresponding to the
output of a single R Markdown story file.
}
\description{
This function knits all R Markdown story files located in a specified directory
and saves the HTML output in another specified directory. It uses
\code{\link{stories_knit_rmd}} to do so.
}
