% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stories.R
\name{stories_knit_rmd}
\alias{stories_knit_rmd}
\title{Render RMarkdown files as HTML with a custom CSS file.}
\usage{
stories_knit_rmd(
  file,
  css_path = here::here("www/styles/sus.css"),
  output_dir = "www/stories/"
)
}
\arguments{
\item{file}{<\code{character}> Full path to the RMarkdown file to be rendered. File
should be have ".Rmd" extension.}

\item{css_path}{<\code{character}> A character string specifying the path to the
CSS file to be applied to the rendered HTML files. Default is "www/sus.css".}

\item{output_dir}{<\code{character}> A character string specifying the directory
where the rendered HTML files will be saved. Default is "www/stories/".}
}
\value{
NULL. The function saves the rendered HTML files to the specified
output directory.
}
\description{
This function renders RMarkdown files as HTML with a custom CSS file. The
rendered HTML files are saved to a specified directory.
}
\seealso{
\code{\link[rmarkdown]{render}} for more details on rendering RMarkdown files.
}
