% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_dictionaries.R
\name{verify_dictionaries}
\alias{verify_dictionaries}
\title{Verify the regions and the scales dictionary}
\usage{
verify_dictionaries(regions_dictionary, scales_dictionary, scales)
}
\arguments{
\item{regions_dictionary}{<\code{data.frame}> The regions dictionary built using
\code{\link[cc.buildr]{regions_dictionary}}}

\item{scales_dictionary}{<\code{data.frame}> The scales dictionary built using
\code{\link[cc.buildr]{build_census_scales}}}

\item{all_tables}{<\verb{named list}> The name of the named list \code{all_tables}
is used to retrieve the shapefiles in the supplied folder.}
}
\value{
Returns nothing if the test passes, or errors if it doesn't.
}
\description{
The use of \code{verify_dictionaries()} is to verify any missing entries in both
dictionaries.
}
